Fit an integer adjusted exponential or gamma distribution

Fit an integer adjusted exponential or gamma distribution

dist_fit(values = NULL, samples = NULL, dist = "exp")

Arguments

values
Numeric vector of values

samples
Numeric, number of samples to take

dist
Character string, which distribution to fit. Defaults to exponential ("exp") but gamma is also supported ("gamma").

Value

A stan fit of an interval censored distribution

Examples

dist_fit(rexp(1:100, 2), samples = 1000, dist = "exp")

#> Inference for Stan model: exp_fit.
#> 2 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=2000.
#> 
#>                mean se_mean   sd   2.5%    25%    50%    75%  97.5% n_eff Rhat
#> lambda         2.07    0.01 0.29   1.56   1.86   2.05   2.24   2.68   629 1.00
#> log_lik[1]    -0.12    0.00 0.04  -0.20  -0.14  -0.11  -0.09  -0.06   628 1.01
#> log_lik[2]    -0.03    0.00 0.02  -0.07  -0.04  -0.03  -0.02  -0.01   616 1.01
#> log_lik[3]    -0.13    0.00 0.04  -0.21  -0.15  -0.12  -0.10  -0.06   629 1.01
#> log_lik[4]    -0.12    0.00 0.04  -0.20  -0.14  -0.11  -0.09  -0.06   628 1.01
#> log_lik[5]    -0.12    0.00 0.04  -0.21  -0.15  -0.12  -0.10  -0.06   628 1.01
#> log_lik[6]    -0.03    0.00 0.01  -0.06  -0.04  -0.03  -0.02  -0.01   613 1.01
#> log_lik[7]    -0.13    0.00 0.04  -0.21  -0.15  -0.12  -0.10  -0.06   629 1.01
#> log_lik[8]    -0.07    0.00 0.03  -0.14  -0.09  -0.07  -0.05  -0.03   624 1.01
#> log_lik[9]    -0.08    0.00 0.03  -0.15  -0.10  -0.08  -0.06  -0.03   625 1.01
#> log_lik[10]   -0.13    0.00 0.04  -0.22  -0.15  -0.12  -0.10  -0.06   629 1.01
#> log_lik[11]   -0.04    0.00 0.02  -0.08  -0.05  -0.03  -0.02  -0.01   616 1.01
#> log_lik[12]   -0.14    0.00 0.04  -0.23  -0.17  -0.14  -0.11  -0.07   629 1.01
#> log_lik[13]   -0.04    0.00 0.02  -0.08  -0.05  -0.03  -0.03  -0.01   617 1.01
#> log_lik[14]   -4.33    0.02 0.59  -5.59  -4.68  -4.28  -3.91  -3.30   629 1.00
#> log_lik[15]   -0.92    0.00 0.12  -1.18  -0.99  -0.91  -0.84  -0.73   629 1.00
#> log_lik[16]   -0.03    0.00 0.02  -0.07  -0.04  -0.03  -0.02  -0.01   615 1.01
#> log_lik[17]   -0.46    0.00 0.05  -0.58  -0.49  -0.45  -0.42  -0.38   627 1.00
#> log_lik[18]   -0.02    0.00 0.01  -0.06  -0.03  -0.02  -0.02  -0.01   611 1.01
#> log_lik[19]   -0.09    0.00 0.03  -0.17  -0.11  -0.09  -0.07  -0.04   626 1.01
#> log_lik[20]   -0.57    0.00 0.07  -0.71  -0.60  -0.56  -0.52  -0.46   628 1.00
#> log_lik[21]   -0.05    0.00 0.02  -0.09  -0.06  -0.04  -0.03  -0.02   619 1.01
#> log_lik[22]   -0.05    0.00 0.02  -0.09  -0.06  -0.04  -0.03  -0.02   619 1.01
#> log_lik[23]   -0.09    0.00 0.03  -0.16  -0.10  -0.08  -0.06  -0.04   626 1.01
#> log_lik[24]   -0.02    0.00 0.01  -0.05  -0.03  -0.02  -0.01  -0.01   609 1.01
#> log_lik[25]   -0.04    0.00 0.02  -0.08  -0.05  -0.03  -0.02  -0.01   616 1.01
#> log_lik[26]   -0.02    0.00 0.01  -0.05  -0.03  -0.02  -0.01  -0.01   610 1.01
#> log_lik[27]   -0.09    0.00 0.03  -0.16  -0.11  -0.09  -0.07  -0.04   626 1.01
#> log_lik[28]   -0.06    0.00 0.02  -0.12  -0.08  -0.06  -0.05  -0.03   623 1.01
#> log_lik[29]   -0.08    0.00 0.03  -0.15  -0.10  -0.08  -0.06  -0.03   625 1.01
#> log_lik[30]   -3.39    0.02 0.46  -4.37  -3.66  -3.35  -3.06  -2.59   629 1.00
#> log_lik[31]   -0.10    0.00 0.03  -0.18  -0.13  -0.10  -0.08  -0.05   627 1.01
#> log_lik[32]   -0.02    0.00 0.01  -0.05  -0.03  -0.02  -0.01  -0.01   609 1.01
#> log_lik[33]   -0.05    0.00 0.02  -0.10  -0.06  -0.05  -0.03  -0.02   620 1.01
#> log_lik[34]   -0.09    0.00 0.03  -0.16  -0.11  -0.09  -0.07  -0.04   626 1.01
#> log_lik[35]   -0.08    0.00 0.03  -0.15  -0.10  -0.08  -0.06  -0.04   625 1.01
#> log_lik[36]   -0.06    0.00 0.02  -0.12  -0.08  -0.06  -0.04  -0.02   622 1.01
#> log_lik[37]   -0.03    0.00 0.02  -0.07  -0.04  -0.03  -0.02  -0.01   615 1.01
#> log_lik[38]   -0.14    0.00 0.04  -0.23  -0.16  -0.13  -0.11  -0.07   629 1.01
#> log_lik[39]   -0.05    0.00 0.02  -0.10  -0.06  -0.05  -0.03  -0.02   620 1.01
#> log_lik[40]   -0.02    0.00 0.01  -0.05  -0.02  -0.02  -0.01   0.00   607 1.01
#> log_lik[41]   -0.09    0.00 0.03  -0.16  -0.11  -0.09  -0.07  -0.04   626 1.01
#> log_lik[42]   -0.10    0.00 0.03  -0.18  -0.13  -0.10  -0.08  -0.05   627 1.01
#> log_lik[43]   -0.04    0.00 0.02  -0.08  -0.05  -0.04  -0.03  -0.01   617 1.01
#> log_lik[44]   -2.34    0.01 0.32  -3.01  -2.52  -2.31  -2.11  -1.80   629 1.00
#> log_lik[45]   -0.02    0.00 0.01  -0.04  -0.02  -0.02  -0.01   0.00   607 1.01
#> log_lik[46]   -0.11    0.00 0.04  -0.19  -0.13  -0.11  -0.09  -0.05   628 1.01
#> log_lik[47]   -0.46    0.00 0.05  -0.58  -0.49  -0.46  -0.43  -0.38   627 1.00
#> log_lik[48]   -0.08    0.00 0.03  -0.14  -0.10  -0.08  -0.06  -0.03   625 1.01
#> log_lik[49]   -0.03    0.00 0.01  -0.06  -0.04  -0.03  -0.02  -0.01   612 1.01
#> log_lik[50]   -0.05    0.00 0.02  -0.10  -0.06  -0.05  -0.03  -0.02   620 1.01
#> log_lik[51]   -0.10    0.00 0.03  -0.18  -0.12  -0.10  -0.08  -0.05   627 1.01
#> log_lik[52]   -0.04    0.00 0.02  -0.09  -0.06  -0.04  -0.03  -0.02   619 1.01
#> log_lik[53]   -0.02    0.00 0.01  -0.05  -0.02  -0.02  -0.01   0.00   607 1.01
#> log_lik[54]   -0.08    0.00 0.03  -0.15  -0.10  -0.08  -0.06  -0.03   625 1.01
#> log_lik[55]   -0.10    0.00 0.03  -0.18  -0.13  -0.10  -0.08  -0.05   627 1.01
#> log_lik[56]   -0.05    0.00 0.02  -0.10  -0.07  -0.05  -0.04  -0.02   621 1.01
#> log_lik[57]   -0.12    0.00 0.04  -0.21  -0.15  -0.12  -0.09  -0.06   628 1.01
#> log_lik[58]   -0.10    0.00 0.03  -0.18  -0.12  -0.10  -0.08  -0.04   627 1.01
#> log_lik[59]   -0.46    0.00 0.05  -0.57  -0.49  -0.45  -0.42  -0.38   627 1.00
#> log_lik[60]   -0.08    0.00 0.03  -0.14  -0.09  -0.07  -0.06  -0.03   625 1.01
#> log_lik[61]   -0.11    0.00 0.04  -0.19  -0.13  -0.11  -0.09  -0.05   628 1.01
#> log_lik[62]   -0.12    0.00 0.01  -0.13  -0.12  -0.12  -0.11  -0.11   539 1.00
#> log_lik[63]   -1.18    0.01 0.15  -1.51  -1.27  -1.17  -1.07  -0.92   629 1.00
#> log_lik[64]   -0.14    0.00 0.04  -0.23  -0.16  -0.13  -0.11  -0.07   629 1.01
#> log_lik[65]   -0.06    0.00 0.02  -0.11  -0.07  -0.05  -0.04  -0.02   622 1.01
#> log_lik[66]   -0.11    0.00 0.04  -0.19  -0.13  -0.10  -0.08  -0.05   628 1.01
#> log_lik[67]   -0.06    0.00 0.02  -0.11  -0.07  -0.05  -0.04  -0.02   621 1.01
#> log_lik[68]   -0.06    0.00 0.02  -0.11  -0.07  -0.05  -0.04  -0.02   622 1.01
#> log_lik[69]   -0.11    0.00 0.04  -0.19  -0.13  -0.10  -0.08  -0.05   627 1.01
#> log_lik[70]   -0.03    0.00 0.01  -0.06  -0.04  -0.03  -0.02  -0.01   613 1.01
#> log_lik[71]   -0.10    0.00 0.03  -0.18  -0.13  -0.10  -0.08  -0.05   627 1.01
#> log_lik[72]   -0.37    0.00 0.04  -0.47  -0.40  -0.37  -0.34  -0.31   626 1.00
#> log_lik[73]   -0.06    0.00 0.02  -0.11  -0.07  -0.05  -0.04  -0.02   621 1.01
#> log_lik[74]   -0.14    0.00 0.04  -0.23  -0.16  -0.13  -0.11  -0.07   629 1.01
#> log_lik[75]   -0.03    0.00 0.01  -0.06  -0.04  -0.03  -0.02  -0.01   613 1.01
#> log_lik[76]   -0.12    0.00 0.04  -0.21  -0.15  -0.12  -0.10  -0.06   628 1.01
#> log_lik[77]   -0.12    0.00 0.04  -0.21  -0.15  -0.12  -0.10  -0.06   628 1.01
#> log_lik[78]   -0.05    0.00 0.02  -0.10  -0.06  -0.05  -0.04  -0.02   621 1.01
#> log_lik[79]   -0.14    0.00 0.04  -0.23  -0.16  -0.13  -0.11  -0.07   629 1.01
#> log_lik[80]   -0.05    0.00 0.02  -0.10  -0.06  -0.04  -0.03  -0.02   620 1.01
#> log_lik[81]   -1.42    0.01 0.19  -1.82  -1.53  -1.40  -1.28  -1.10   629 1.00
#> log_lik[82]   -1.07    0.01 0.14  -1.37  -1.16  -1.06  -0.98  -0.84   629 1.00
#> log_lik[83]   -0.12    0.00 0.04  -0.20  -0.14  -0.11  -0.09  -0.06   628 1.01
#> log_lik[84]   -0.10    0.00 0.03  -0.17  -0.12  -0.09  -0.07  -0.04   627 1.01
#> log_lik[85]   -0.73    0.00 0.09  -0.93  -0.79  -0.73  -0.67  -0.59   628 1.00
#> log_lik[86]   -0.09    0.00 0.03  -0.15  -0.10  -0.08  -0.06  -0.04   626 1.01
#> log_lik[87]   -0.02    0.00 0.01  -0.05  -0.03  -0.02  -0.01  -0.01   609 1.01
#> log_lik[88]   -0.08    0.00 0.03  -0.15  -0.10  -0.08  -0.06  -0.04   625 1.01
#> log_lik[89]   -0.12    0.00 0.04  -0.20  -0.14  -0.11  -0.09  -0.06   628 1.01
#> log_lik[90]   -0.05    0.00 0.02  -0.11  -0.07  -0.05  -0.04  -0.02   621 1.01
#> log_lik[91]   -0.03    0.00 0.02  -0.07  -0.04  -0.03  -0.02  -0.01   616 1.01
#> log_lik[92]   -0.10    0.00 0.03  -0.18  -0.12  -0.10  -0.08  -0.05   627 1.01
#> log_lik[93]   -0.13    0.00 0.04  -0.23  -0.16  -0.13  -0.11  -0.07   629 1.01
#> log_lik[94]   -0.68    0.00 0.08  -0.86  -0.73  -0.67  -0.62  -0.54   628 1.00
#> log_lik[95]   -0.06    0.00 0.02  -0.11  -0.07  -0.06  -0.04  -0.02   622 1.01
#> log_lik[96]   -0.03    0.00 0.02  -0.07  -0.04  -0.03  -0.02  -0.01   616 1.01
#> log_lik[97]   -0.04    0.00 0.02  -0.08  -0.05  -0.03  -0.02  -0.01   617 1.01
#> log_lik[98]   -0.06    0.00 0.02  -0.12  -0.08  -0.06  -0.05  -0.02   623 1.01
#> log_lik[99]   -0.02    0.00 0.01  -0.05  -0.03  -0.02  -0.01   0.00   607 1.01
#> log_lik[100]  -0.04    0.00 0.02  -0.08  -0.05  -0.03  -0.02  -0.01   617 1.01
#> lp__         -24.05    0.03 0.69 -26.00 -24.22 -23.78 -23.61 -23.57   464 1.00
#> 
#> Samples were drawn using NUTS(diag_e) at Mon Jun  1 12:33:46 2020.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).