Estimate time-varying measures and forecast

Estimate time-varying measures and forecast

epi_measures_pipeline(
  nowcast = NULL,
  generation_times = NULL,
  min_est_date = NULL,
  gt_samples = 1,
  rt_samples = 5,
  rt_windows = 7,
  rate_window = 7,
  rt_prior = NULL,
  forecast_model = NULL,
  horizon = 0,
  verbose = TRUE
)

Arguments

nowcast
A nowcast as produced by nowcast_pipeline

generation_times
A matrix with columns representing samples and rows representing the probability of the generation timebeing on that day.

min_est_date
Date to begin estimation.

gt_samples
Numeric, the number of samples to take from the generaiton times supplied

rt_samples
Numeric, the number of samples to take from the estimated R distribution for each time point.

rt_windows
Numeric vector, windows over which to estimate time-varying R. The best performing window will be selected per serial interval sample by default (based on which window best forecasts current cases).

rate_window
Numeric, the window to use to estimate the rate of spread.

rt_prior
A list defining the reproduction number prior containing the mean (mean_prior) and standard deviation (std_prior)

forecast_model
An uninitialised bsts model passed to EpiSoon::forecast_rt to be used for forecasting future Rt values. An example of the required structure is: function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}.

horizon
Numeric, defaults to 0. The horizon over which to forecast Rts and cases.

verbose
Logical, defaults to TRUE. Should progress messages be shown.

Value

A list of data frames containing reproduction number estimates, case forecasts, rate of growth estimates both summarised and raw.

Examples

## Construct example distributions
## reporting delay dist
delay_dist <- EpiNow::lognorm_dist_def(mean = 3,
                                      mean_sd = 1,
                                      sd = 3,
                                      sd_sd = 1,
                                      max_value = 30,
                                      samples = 1)

## incubation delay dist
incubation_dist <- delay_dist

## Uses example case vector from EpiSoon
cases <- data.table::setDT(EpiSoon::example_obs_cases)
cases <- cases[, `:=`(confirm = as.integer(cases), import_status = "local")]

## Basic nowcast
nowcast <- nowcast_pipeline(reported_cases = cases,
                            target_date = max(cases$date),
                            delay_defs = delay_dist,
                            incubation_defs = incubation_dist)

## Estimate parameters                           
estimates <- epi_measures_pipeline(nowcast[type %in% "infection_upscaled"],
                       generation_times = EpiNow::covid_generation_times,
                       rt_prior = list(mean_prior = 2.6, std_prior = 2))

#> Estimate time-varying R0

#> Summarising time-varying R0

#> Summarising forecast cases

#> Estimate time-varying rate of growth

#> 
#> Attaching package: ‘purrr’

#> The following object is masked from ‘package:data.table’:
#> 
#>     transpose

estimates

#> $R0
#>                   type       date rt_type    bottom      top     lower    upper
#>  1: infection_upscaled 2020-01-17 nowcast 0.7466803 3.006430 0.7466803 1.240275
#>  2: infection_upscaled 2020-01-18 nowcast 0.8864384 3.570888 1.2786435 1.841273
#>  3: infection_upscaled 2020-01-19 nowcast 2.0091058 4.437581 2.0091058 2.710940
#>  4: infection_upscaled 2020-01-20 nowcast 1.6018509 2.256486 1.7546234 1.955850
#>  5: infection_upscaled 2020-01-21 nowcast 2.0716867 2.665287 2.6230063 2.665287
#>  6: infection_upscaled 2020-01-22 nowcast 1.6274709 2.909955 2.3392161 2.909955
#>  7: infection_upscaled 2020-01-23 nowcast 1.3913986 2.062623 1.4779104 1.611343
#>  8: infection_upscaled 2020-01-24 nowcast 1.0403333 2.029404 1.4661985 1.677742
#>  9: infection_upscaled 2020-01-25 nowcast 1.0330824 2.445151 1.0330824 1.485077
#> 10: infection_upscaled 2020-01-26 nowcast 1.4280489 2.151118 1.7192008 1.940777
#> 11: infection_upscaled 2020-01-27 nowcast 1.4031655 1.980483 1.4031655 1.522599
#> 12: infection_upscaled 2020-01-28 nowcast 1.3629897 1.876518 1.3629897 1.493564
#> 13: infection_upscaled 2020-01-29 nowcast 1.5291823 2.007183 1.7123566 1.794109
#> 14: infection_upscaled 2020-01-30 nowcast 1.2214383 1.419965 1.3249586 1.419965
#> 15: infection_upscaled 2020-01-31 nowcast 1.4692658 2.031425 1.4692658 1.697421
#> 16: infection_upscaled 2020-02-01 nowcast 1.2889701 1.565084 1.2889701 1.376827
#> 17: infection_upscaled 2020-02-02 nowcast 1.2676366 1.893551 1.7281252 1.893551
#> 18: infection_upscaled 2020-02-03 nowcast 1.0180099 1.568886 1.4990798 1.568886
#> 19: infection_upscaled 2020-02-04 nowcast 1.3427592 1.617291 1.5799463 1.594510
#> 20: infection_upscaled 2020-02-05 nowcast 1.2208566 1.553431 1.2208566 1.246987
#> 21: infection_upscaled 2020-02-06 nowcast 1.0756814 1.423310 1.1895265 1.293894
#> 22: infection_upscaled 2020-02-07 nowcast 1.1511530 1.721054 1.3177636 1.408595
#> 23: infection_upscaled 2020-02-08 nowcast 1.2907245 1.698699 1.5179218 1.698699
#> 24: infection_upscaled 2020-02-09 nowcast 1.3360168 1.710270 1.4931149 1.528359
#> 25: infection_upscaled 2020-02-10 nowcast 1.3213034 1.585728 1.3213034 1.436011
#> 26: infection_upscaled 2020-02-11 nowcast 1.4218020 1.685952 1.4218020 1.437369
#> 27: infection_upscaled 2020-02-12 nowcast 1.3194175 1.672174 1.3776787 1.441203
#> 28: infection_upscaled 2020-02-13 nowcast 1.2139617 1.525500 1.2139617 1.302122
#> 29: infection_upscaled 2020-02-14 nowcast 1.4193496 1.667752 1.4499892 1.533110
#> 30: infection_upscaled 2020-02-15 nowcast 1.3594522 1.672884 1.6330792 1.672884
#> 31: infection_upscaled 2020-02-16 nowcast 1.2845592 1.545096 1.4612553 1.545096
#> 32: infection_upscaled 2020-02-17 nowcast 1.4469437 1.670705 1.4469437 1.535656
#> 33: infection_upscaled 2020-02-18 nowcast 1.3969442 1.546580 1.3969442 1.421942
#> 34: infection_upscaled 2020-02-19 nowcast 1.3771659 1.522822 1.3771659 1.412510
#> 35: infection_upscaled 2020-02-20 nowcast 1.3748343 1.559742 1.4304941 1.459397
#> 36: infection_upscaled 2020-02-21 nowcast 1.2481871 1.532993 1.2481871 1.381135
#> 37: infection_upscaled 2020-02-22 nowcast 1.2735373 1.431546 1.4149205 1.431546
#> 38: infection_upscaled 2020-02-23 nowcast 1.3051997 1.456052 1.3439783 1.402537
#> 39: infection_upscaled 2020-02-24 nowcast 1.2356516 1.457084 1.3015738 1.368422
#> 40: infection_upscaled 2020-02-25 nowcast 1.2166608 1.511916 1.2789867 1.286997
#> 41: infection_upscaled 2020-02-26 nowcast 1.1714963 1.271753 1.2585896 1.271753
#> 42: infection_upscaled 2020-02-27 nowcast 1.1991284 1.256669 1.1991284 1.208677
#> 43: infection_upscaled 2020-02-28 nowcast 1.1221917 1.215119 1.1727447 1.198325
#> 44: infection_upscaled 2020-02-29 nowcast 1.1555807 1.223686 1.1651687 1.193660
#> 45: infection_upscaled 2020-03-01 nowcast 1.0829161 1.228728 1.1026919 1.153561
#> 46: infection_upscaled 2020-03-02 nowcast 1.0495864 1.188871 1.1007279 1.126226
#> 47: infection_upscaled 2020-03-03 nowcast 1.1465917 1.229056 1.1465917 1.186643
#> 48: infection_upscaled 2020-03-04 nowcast 1.0912553 1.174296 1.0912553 1.107252
#> 49: infection_upscaled 2020-03-05 nowcast 1.1359939 1.274421 1.1359939 1.188924
#> 50: infection_upscaled 2020-03-06 nowcast 1.1037206 1.211454 1.2035683 1.211454
#> 51: infection_upscaled 2020-03-07 nowcast 1.1741737 1.262313 1.1741737 1.206635
#> 52: infection_upscaled 2020-03-08 nowcast 1.2207433 1.297691 1.2207433 1.241641
#> 53: infection_upscaled 2020-03-09 nowcast 1.1717427 1.282012 1.1717427 1.191974
#> 54: infection_upscaled 2020-03-10 nowcast 1.1255103 1.209293 1.1824386 1.209293
#> 55: infection_upscaled 2020-03-11 nowcast 1.1772820 1.229559 1.1772820 1.184618
#> 56: infection_upscaled 2020-03-12 nowcast 1.1607926 1.233918 1.1974317 1.233918
#> 57: infection_upscaled 2020-03-13 nowcast 1.1449609 1.205278 1.1792924 1.205278
#> 58: infection_upscaled 2020-03-14 nowcast 1.0865500 1.183295 1.1507180 1.183295
#>                   type       date rt_type    bottom      top     lower    upper
#>       median     mean        std prob_control mean_window sd_window mean_crps
#>  1: 1.240275 1.669805 0.90369888          0.2           7         0      0.16
#>  2: 1.714775 1.858403 1.02887841          0.2           7         0      0.80
#>  3: 2.710940 2.913762 1.01365622          0.0           7         0      0.76
#>  4: 1.860799 1.885922 0.24534366          0.0           7         0      1.76
#>  5: 2.623006 2.435070 0.28389486          0.0           7         0      1.64
#>  6: 2.339216 2.288826 0.53616999          0.0           7         0      1.04
#>  7: 1.546736 1.618002 0.26162254          0.0           7         0      0.24
#>  8: 1.545917 1.551919 0.35809454          0.0           7         0      0.56
#>  9: 1.485077 1.730812 0.62356984          0.0           7         0      0.60
#> 10: 1.828216 1.813472 0.26826605          0.0           7         0      0.60
#> 11: 1.522599 1.593823 0.23941284          0.0           7         0      0.52
#> 12: 1.493564 1.576664 0.22468564          0.0           7         0      1.44
#> 13: 1.730944 1.754755 0.17213087          0.0           7         0      1.32
#> 14: 1.324959 1.310754 0.08278257          0.0           7         0      1.44
#> 15: 1.697421 1.714296 0.23008647          0.0           7         0      2.16
#> 16: 1.376827 1.412250 0.12619517          0.0           7         0      5.24
#> 17: 1.728125 1.625767 0.29216263          0.0           7         0      5.04
#> 18: 1.499080 1.387094 0.23598305          0.0           7         0      1.60
#> 19: 1.591189 1.545139 0.11394425          0.0           7         0      2.72
#> 20: 1.246987 1.306783 0.13905364          0.0           7         0      4.24
#> 21: 1.245979 1.245678 0.12838397          0.0           7         0      2.76
#> 22: 1.374859 1.394685 0.20756200          0.0           7         0      6.24
#> 23: 1.517922 1.471395 0.16829129          0.0           7         0      4.84
#> 24: 1.517719 1.517096 0.13304925          0.0           7         0      1.16
#> 25: 1.436011 1.461160 0.12011220          0.0           7         0      3.28
#> 26: 1.437369 1.520364 0.12462512          0.0           7         0      3.00
#> 27: 1.415786 1.445252 0.13488750          0.0           7         0      1.56
#> 28: 1.302122 1.364947 0.13973444          0.0           7         0      4.12
#> 29: 1.505450 1.515130 0.09634289          0.0           7         0      1.52
#> 30: 1.633079 1.572523 0.12868598          0.0           7         0      7.36
#> 31: 1.461255 1.427764 0.10760312          0.0           7         0      1.28
#> 32: 1.535656 1.560001 0.09325059          0.0           7         0      3.12
#> 33: 1.421942 1.446045 0.06141216          0.0           7         0      6.52
#> 34: 1.412510 1.427873 0.05710594          0.0           7         0      9.24
#> 35: 1.448473 1.454588 0.06719184          0.0           7         0      8.68
#> 36: 1.381135 1.377119 0.11442665          0.0           7         0      2.44
#> 37: 1.414920 1.369454 0.07173354          0.0           7         0      3.48
#> 38: 1.391011 1.379756 0.05767456          0.0           7         0      5.16
#> 39: 1.349954 1.342537 0.08208016          0.0           7         0      8.68
#> 40: 1.279541 1.314820 0.11378120          0.0           7         0      3.96
#> 41: 1.258590 1.231752 0.04653059          0.0           7         0     28.16
#> 42: 1.208677 1.222396 0.02500451          0.0           7         0      4.40
#> 43: 1.187628 1.179202 0.03542102          0.0           7         0     25.32
#> 44: 1.193622 1.186344 0.02690593          0.0           7         0      3.68
#> 45: 1.140076 1.141595 0.05633461          0.0           7         0      2.84
#> 46: 1.121877 1.117458 0.05020618          0.0           7         0     23.16
#> 47: 1.186643 1.186139 0.03559976          0.0           7         0      1.88
#> 48: 1.107252 1.119692 0.03489331          0.0           7         0      2.32
#> 49: 1.188924 1.193717 0.05573907          0.0           7         0      3.84
#> 50: 1.203568 1.179519 0.04517022          0.0           7         0     11.20
#> 51: 1.206635 1.213581 0.03950063          0.0           7         0     14.16
#> 52: 1.241641 1.254974 0.03187404          0.0           7         0      3.72
#> 53: 1.191974 1.217578 0.05183788          0.0           7         0     13.96
#> 54: 1.182439 1.167300 0.03659044          0.0           7         0      6.00
#> 55: 1.184618 1.194927 0.02234939          0.0           7         0      4.96
#> 56: 1.197432 1.196733 0.03027523          0.0           7         0      9.08
#> 57: 1.179292 1.176838 0.02546946          0.0           7         0      4.64
#> 58: 1.150718 1.142070 0.03762434          0.0           7         0      1.36
#>       median     mean        std prob_control mean_window sd_window mean_crps
#>     sd_crps R0_range
#>  1:       0   <list>
#>  2:       0   <list>
#>  3:       0   <list>
#>  4:       0   <list>
#>  5:       0   <list>
#>  6:       0   <list>
#>  7:       0   <list>
#>  8:       0   <list>
#>  9:       0   <list>
#> 10:       0   <list>
#> 11:       0   <list>
#> 12:       0   <list>
#> 13:       0   <list>
#> 14:       0   <list>
#> 15:       0   <list>
#> 16:       0   <list>
#> 17:       0   <list>
#> 18:       0   <list>
#> 19:       0   <list>
#> 20:       0   <list>
#> 21:       0   <list>
#> 22:       0   <list>
#> 23:       0   <list>
#> 24:       0   <list>
#> 25:       0   <list>
#> 26:       0   <list>
#> 27:       0   <list>
#> 28:       0   <list>
#> 29:       0   <list>
#> 30:       0   <list>
#> 31:       0   <list>
#> 32:       0   <list>
#> 33:       0   <list>
#> 34:       0   <list>
#> 35:       0   <list>
#> 36:       0   <list>
#> 37:       0   <list>
#> 38:       0   <list>
#> 39:       0   <list>
#> 40:       0   <list>
#> 41:       0   <list>
#> 42:       0   <list>
#> 43:       0   <list>
#> 44:       0   <list>
#> 45:       0   <list>
#> 46:       0   <list>
#> 47:       0   <list>
#> 48:       0   <list>
#> 49:       0   <list>
#> 50:       0   <list>
#> 51:       0   <list>
#> 52:       0   <list>
#> 53:       0   <list>
#> 54:       0   <list>
#> 55:       0   <list>
#> 56:       0   <list>
#> 57:       0   <list>
#> 58:       0   <list>
#>     sd_crps R0_range
#> 
#> $rate_of_spread
#>                  type overall_little_r time_varying_r
#> 1: infection_upscaled           <list>   <data.table>
#> 
#> $raw_R0
#>            date         R sample crps window rt_type               type
#>   1: 2020-01-17 0.7466803      1 0.16      7 nowcast infection_upscaled
#>   2: 2020-01-17 1.2070038      1 0.16      7 nowcast infection_upscaled
#>   3: 2020-01-17 1.2402748      1 0.16      7 nowcast infection_upscaled
#>   4: 2020-01-17 2.1486380      1 0.16      7 nowcast infection_upscaled
#>   5: 2020-01-17 3.0064296      1 0.16      7 nowcast infection_upscaled
#>  ---                                                                   
#> 286: 2020-03-14 1.0865500      1 1.36      7 nowcast infection_upscaled
#> 287: 2020-03-14 1.1250233      1 1.36      7 nowcast infection_upscaled
#> 288: 2020-03-14 1.1507180      1 1.36      7 nowcast infection_upscaled
#> 289: 2020-03-14 1.1647643      1 1.36      7 nowcast infection_upscaled
#> 290: 2020-03-14 1.1832948      1 1.36      7 nowcast infection_upscaled
#>