Estimate the time varying R0 - using EpiEstim

Estimate the time varying R0 - using EpiEstim

estimate_R0(
  cases = NULL,
  generation_times = NULL,
  rt_prior = NULL,
  windows = NULL,
  gt_samples = 100,
  rt_samples = 100,
  min_est_date = NULL,
  forecast_model = NULL,
  horizon = 0
)

Arguments

cases
A dataframe containing a list of local cases with the following variables: date, cases, and import_status

generation_times
A matrix with columns representing samples and rows representing the probability of the generation timebeing on that day.

rt_prior
A list defining the reproduction number prior containing the mean (mean_prior) and standard deviation (std_prior)

windows
Numeric vector, windows over which to estimate time-varying R. The best performing window will be selected per serial interval sample by default (based on which window best forecasts current cases).

gt_samples
Numeric, the number of samples to take from the generaiton times supplied

rt_samples
Numeric, the number of samples to take from the estimated R distribution for each time point.

min_est_date
Date to begin estimation.

forecast_model
An uninitialised bsts model passed to EpiSoon::forecast_rt to be used for forecasting future Rt values. An example of the required structure is: function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}.

horizon
Numeric, defaults to 0. The horizon over which to forecast Rts and cases.

Value

A tibble containing the date and summarised R estimte.

Examples

## Nowcast Rts                  
estimates <- estimate_R0(cases = EpiSoon::example_obs_cases,
                         generation_times = as.matrix(EpiNow::covid_generation_times[,2]),
                         rt_prior = list(mean_prior = 2.6, std_prior = 2),
                         windows = c(1, 3, 7), rt_samples = 10, gt_samples = 2,
                         min_est_date =  as.Date("2020-02-18"))

estimates$rts

#>            date         R sample   crps window rt_type
#>   1: 2020-02-18 0.8660468      1   0.21      3 nowcast
#>   2: 2020-02-18 1.1586639      2   0.21      3 nowcast
#>   3: 2020-02-18 1.1777688      3   0.21      3 nowcast
#>   4: 2020-02-18 1.4778703      4   0.21      3 nowcast
#>   5: 2020-02-18 1.8766432      5   0.21      3 nowcast
#>  ---                                                  
#> 686: 2020-03-22 1.3970775     12 165.58      1 nowcast
#> 687: 2020-03-22 1.3995913     14 165.58      1 nowcast
#> 688: 2020-03-22 1.4154107     16 165.58      1 nowcast
#> 689: 2020-03-22 1.4272942     18 165.58      1 nowcast
#> 690: 2020-03-22 1.4550747     20 165.58      1 nowcast

## Nowcast Rts, forecast Rts and the forecast cases
estimates <- estimate_R0(cases = EpiSoon::example_obs_cases,
                         generation_times = as.matrix(EpiNow::covid_generation_times[,1]),
                         rt_prior = list(mean_prior = 2.6, std_prior = 2),
                         windows = c(1, 3, 7), rt_samples = 10, gt_samples = 2,
                         min_est_date =  as.Date("2020-02-18"),
                         forecast_model = function(...){EpiSoon::fable_model(model = fable::ETS(y ~ trend("A")), ...)},
                         horizon = 7)

## Rt estimates and forecasts
estimates$rts

#>            date         R sample crps window  rt_type
#>   1: 2020-02-18 0.9172282      1 0.25      7  nowcast
#>   2: 2020-02-18 1.0822154      2 0.25      7  nowcast
#>   3: 2020-02-18 1.2976365      3 0.25      7  nowcast
#>   4: 2020-02-18 1.3400988      4 0.25      7  nowcast
#>   5: 2020-02-18 1.3567841      5 0.25      7  nowcast
#>  ---                                                 
#> 816: 2020-03-29 1.6563804     12   NA     NA forecast
#> 817: 2020-03-29 2.1935595     14   NA     NA forecast
#> 818: 2020-03-29 1.4503716     16   NA     NA forecast
#> 819: 2020-03-29 1.3547133     18   NA     NA forecast
#> 820: 2020-03-29 1.5227681     20   NA     NA forecast

## Case forecasts
estimates$cases

#>      sample       date cases horizon  rt_type
#>   1:      1 2020-03-23   562       1 forecast
#>   2:      1 2020-03-24   678       2 forecast
#>   3:      1 2020-03-25   681       3 forecast
#>   4:      1 2020-03-26   656       4 forecast
#>   5:      1 2020-03-27   769       5 forecast
#>  ---                                         
#> 136:     20 2020-03-25   609       3 forecast
#> 137:     20 2020-03-26   898       4 forecast
#> 138:     20 2020-03-27   922       5 forecast
#> 139:     20 2020-03-28  1689       6 forecast
#> 140:     20 2020-03-29  1264       7 forecast