Regional Realtime Pipeline

Runs a national realtime pipeline followed by a pipeline for each individual region.

regional_rt_pipeline(
  cases = NULL,
  linelist = NULL,
  target_folder = "results",
  regional_delay = FALSE,
  merge_onsets = FALSE,
  case_limit = 40,
  onset_modifier = NULL,
  bootstraps = 1,
  bootstrap_samples = 100,
  regions_in_parallel = TRUE,
  verbose = FALSE,
  samples = 1000,
  ...
)

Arguments

cases
A dataframe of cases (confirm) by date of confirmation (date), import status (import_status; ("imp)), and region (region).

linelist
A dataframe of of cases (by row) containing the following variables: import_status (values "local" and "imported"), date_onset, date_confirm, report_delay, and region. If a national linelist is not available a proxy linelist may be used but in this case merge_onsets should be set to FALSE.

target_folder
Character string, name of the folder in which to save the results.

regional_delay
Logical defaults to FALSE. Should reporting delays be estimated by region.

merge_onsets
Logical defaults to FALSE. Should available onset data be used. Typically if regional_delay is

case_limit
Numeric, the minimum number of cases in a region required for that region to be evaluated. Defaults to 10. set to FALSE this should also be FALSE

onset_modifier
data.frame containing a date variable and a function modifier variable. This is used to modify estimated cases by onset date. modifier must be a function that returns a proportion when called (enables inclusion of uncertainty) and takes the following arguments: n (samples to return) and status ("local" or "import").

regions_in_parallel
Logical, should regions be run in parallel or sequentially (allowing for) within pipeline parallisation. Defaults to TRUE.

verbose
Logical, defaults to FALSE. Should progress messages be shown for each reigon?

samples
Numeric, the number of pseudo linelists to generate. Defaults to 1000.

...

Examples

## Code
regional_rt_pipeline

#> function (cases = NULL, linelist = NULL, target_folder = "results", 
#>     regional_delay = FALSE, merge_onsets = FALSE, case_limit = 40, 
#>     onset_modifier = NULL, bootstraps = 1, bootstrap_samples = 100, 
#>     regions_in_parallel = TRUE, verbose = FALSE, samples = 1000, 
#>     ...) 
#> {
#>     cases <- data.table::as.data.table(cases)
#>     linelist <- data.table::as.data.table(linelist)
#>     target_date <- as.character(max(cases$date))
#>     message("Running pipeline for ", target_date)
#>     eval_regions <- data.table::copy(cases)[.(cases = sum(cases, 
#>         na.rm = TRUE)), by = c("region", "date")][cases >= case_limit]$region
#>     eval_regions <- unique(eval_regions)
#>     cases <- cases[!is.na(region)][region %in% eval_regions]
#>     message("Running the pipeline for: ", paste(eval_regions, 
#>         collapse = ", "))
#>     cases_grid <- cases[, .(date = seq(min(date), max(date), 
#>         by = "days"), import_status = list(list("local", "imported"))), 
#>         by = "region"][, .(import_status = unlist(import_status)), 
#>         by = c("date", "region")]
#>     cases <- cases[cases_grid, on = c("date", "region", "import_status")][is.na(cases), 
#>         `:=`(cases, 0)]
#>     cases <- data.table::setorder(cases, region, import_status, 
#>         date)
#>     regions <- unique(cases$region)
#>     if (!regional_delay) {
#>         message("Using a national linelist so not merging onsets and fitting a single reporting delay")
#>         merge_onsets <- FALSE
#>         report_delay_fns <- EpiNow::get_delay_sample_fn(linelist = linelist[, 
#>             `:=`(delay_confirmatin, report_delay)], samples = samples, 
#>             bootstraps = bootstraps, bootstrap_samples = bootstrap_samples)
#>     }
#>     else {
#>         report_delay_fns <- NULL
#>     }
#>     message("Running pipelines by region")
#>     region_rt <- purrr::partial(EpiNow::rt_pipeline, target_date = target_date, 
#>         merge_actual_onsets = merge_onsets, samples = samples, 
#>         report_delay_fns = report_delay_fns, verbose = verbose, 
#>         ...)
#>     run_region <- function(target_region) {
#>         message("Running Rt pipeline for ", target_region)
#>         regional_cases <- cases[regions %in% target_region]
#>         if (regional_delay) {
#>             regional_linelist <- linelist[region %in% target_region]
#>         }
#>         else {
#>             regional_linelist <- linelist
#>         }
#>         if (!is.null(onset_modifier)) {
#>             region_onset_modifier <- data.table::setDT(onset_modifier)[region %in% 
#>                 target_region]
#>             region_onset_modifier <- region_onset_modifier[, 
#>                 `:=`(region, NULL)]
#>         }
#>         else {
#>             region_onset_modifier <- NULL
#>         }
#>         region_rt(cases = regional_cases, linelist = regional_linelist, 
#>             onset_modifier = region_onset_modifier, target_folder = file.path(target_folder, 
#>                 target_region))
#>         rm(list = ls())
#>         return(invisible(NULL))
#>     }
#>     if (regions_in_parallel) {
#>         out <- furrr::future_map(regions, run_region, .progress = TRUE, 
#>             .options = furrr::future_options(scheduling = Inf, 
#>                 globals = c("region_rt", "target_date", "merge_onsets", 
#>                   "samples", "report_delay_fns", "verbose", "cases", 
#>                   "linelist", "onset_modifier", "regions", "target_folder")))
#>     }
#>     else {
#>         out <- purrr::map(regions, run_region)
#>     }
#>     return(invisible(NULL))
#> }
#> <bytecode: 0x559f7f66b3e8>
#> <environment: namespace:EpiNow>