Summarise Realtime Results

Summarise Realtime Results

summarise_results(
  regions = NULL,
  results_dir = "results",
  target_date = NULL,
  region_scale = "Region"
)

Arguments

regions
A character string containing the list of regions to extract results for (must all have results for the same target date).

results_dir
A character string indicating the location of the results directory to extract results from.

target_date
A character string indicating the target date to extract results for. All regions must have results for this date.

region_scale
A character string indicating the name to give the regions being summarised.

Value

Examples

## Code

summarise_results

#> function (regions = NULL, results_dir = "results", target_date = NULL, 
#>     region_scale = "Region") 
#> {
#>     load_data <- purrr::partial(load_nowcast_result, date = target_date, 
#>         result_dir = results_dir)
#>     estimates <- data.table::data.table(region = names(regions), 
#>         `New confirmed cases by infection date` = purrr::map(regions, 
#>             ~load_data("current_cases.rds", .)), `Expected change in daily cases` = map_prob_change(purrr::map_dbl(regions, 
#>             ~load_data("prob_control_latest.rds", .))), `Effective reproduction no.` = purrr::map(regions, 
#>             ~load_data("bigr_eff_latest.rds", .)), `Doubling/halving time (days)` = purrr::map_chr(regions, 
#>             ~load_data("doubling_time_latest.rds", .)))
#>     numeric_estimates <- estimates[, .(region, `New confirmed cases by infection date`, 
#>         `Effective reproduction no.`, `Expected change in daily cases`)]
#>     numeric_estimates <- data.table::melt(numeric_estimates, 
#>         measure.vars = c("New confirmed cases by infection date", 
#>             "Effective reproduction no."), variable.name = "metric", 
#>         value.name = "value")
#>     numeric_estimates <- numeric_estimates[, `:=`(lower = purrr::map_dbl(value, 
#>         ~.[[1]]$lower), upper = purrr::map_dbl(value, ~.[[1]]$upper), 
#>         mid_lower = purrr::map_dbl(value, ~.[[1]]$mid_lower), 
#>         mid_upper = purrr::map_dbl(value, ~.[[1]]$mid_upper))][, 
#>         `:=`(metric, factor(metric, levels = c("New confirmed cases by infection date", 
#>             "Effective reproduction no.")))]
#>     high_inc_regions <- unique(data.table::setorderv(numeric_estimates, 
#>         cols = "upper", order = -1)$region)
#>     numeric_estimates <- numeric_estimates[, `:=`(region, factor(region, 
#>         levels = high_inc_regions))]
#>     estimates <- estimates[, `:=`(`New confirmed cases by infection date` = EpiNow::make_conf(purrr::map(`New confirmed cases by infection date`, 
#>         ~.[[1]]), digits = 0), `Effective reproduction no.` = EpiNow::make_conf(purrr::map(`Effective reproduction no.`, 
#>         ~.[[1]]), digits = 1))]
#>     estimates <- estimates[, `:=`((region_scale), region)][, 
#>         `:=`(region, NULL)]
#>     estimates <- estimates[, c((region_scale), colnames(estimates)[-ncol(estimates)]), 
#>         with = FALSE]
#>     out <- list(estimates, numeric_estimates, high_inc_regions)
#>     names(out) <- c("table", "data", "regions_by_inc")
#>     return(out)
#> }
#> <bytecode: 0x7fdcdabf0698>
#> <environment: namespace:EpiNow>