Adjust from Case Counts by Infection Date to Date of Report

Stochastic mapping from cases by date of infection to date of report via date of onset. Essentially reversal of nowcast_pipeline.

adjust_infection_to_report(
  infections,
  delay_def,
  incubation_def,
  reporting_effect,
  type = "sample",
  return_onset = FALSE
)

Arguments

infections
data.table containing a date variable and a numeric cases variable.

delay_def
A single row data.table that defines the delay distribution (model, parameters and maximum delay for each model). See lognorm_dist_def for an example of the structure.

incubation_def
A single row data.table that defines the incubation distribution (model, parameters and maximum delay for each model). See lognorm_dist_def for an example of the structure.

reporting_effect
A numeric vector of length 7 that allows the scaling of reported cases by the day on which they report (1 = Monday, 7 = Sunday). By default no scaling occurs.

type
Character string indicating the method to use to transfrom counts. Supports either "sample" which approximates sampling or "median" would shift by the median of the distribution.

return_onset
Logical, defaults to FALSE. Should cases by date of onset also be returned?

Value

A data.table containing a date variable (date of report) and a cases variable. If return_onset = TRUE there will be a third variable reference which indicates what the date variable refers to.

Examples

## Define example cases
cases <- data.table::as.data.table(EpiSoon::example_obs_cases)

cases <- cases[, `:=`(confirm = as.integer(cases), import_status = "local")]

## Define a single report delay distribution
delay_def <- EpiNow::lognorm_dist_def(mean = 5,
                                      mean_sd = 1,
                                      sd = 3,
                                      sd_sd = 1,
                                      max_value = 30,
                                      samples = 1)

## Define a single incubation period
incubation_def <- EpiNow::lognorm_dist_def(mean = EpiNow::covid_incubation_period[1, ]$mean,
                                           mean_sd = EpiNow::covid_incubation_period[1, ]$mean_sd,
                                           sd = EpiNow::covid_incubation_period[1, ]$sd,
                                           sd_sd = EpiNow::covid_incubation_period[1, ]$sd_sd,
                                           max_value = 30, samples = 1)

## Perform a nowcast
nowcast <- nowcast_pipeline(reported_cases = cases,
                            target_date = max(cases$date),
                            delay_defs = delay_def,
                            incubation_defs = incubation_def)

#> New names:
#> * NA -> ...1
#> * NA -> ...2
#> * NA -> ...3
#> * NA -> ...4
#> * NA -> ...5
#> * ...

infections <- nowcast[type %in% "infection_upscaled" & import_status %in% "local"]
infections <- infections[, `:=`(type = NULL, import_status = NULL)]

## Simple mapping
report <- adjust_infection_to_report(infections, delay_def, incubation_def)

#> New names:
#> * NA -> ...1
#> * NA -> ...2
#> * NA -> ...3
#> * NA -> ...4
#> * NA -> ...5
#> * ...

#> New names:
#> * NA -> ...1
#> * NA -> ...2
#> * NA -> ...3
#> * NA -> ...4
#> * NA -> ...5
#> * ...

print(report)

#>           date cases
#>  1: 2020-01-18     1
#>  2: 2020-01-19     0
#>  3: 2020-01-20     1
#>  4: 2020-01-21     0
#>  5: 2020-01-22     0
#>  6: 2020-01-23     0
#>  7: 2020-01-24     0
#>  8: 2020-01-25     0
#>  9: 2020-01-26     0
#> 10: 2020-01-27     0
#> 11: 2020-01-28     0
#> 12: 2020-01-29     2
#> 13: 2020-01-30     1
#> 14: 2020-01-31     1
#> 15: 2020-02-01     0
#> 16: 2020-02-02     0
#> 17: 2020-02-03     1
#> 18: 2020-02-04     1
#> 19: 2020-02-05     4
#> 20: 2020-02-06     3
#> 21: 2020-02-07     2
#> 22: 2020-02-08     0
#> 23: 2020-02-09     0
#> 24: 2020-02-10     2
#> 25: 2020-02-11     1
#> 26: 2020-02-12     1
#> 27: 2020-02-13     2
#> 28: 2020-02-14     5
#> 29: 2020-02-15     9
#> 30: 2020-02-16    10
#> 31: 2020-02-17    19
#> 32: 2020-02-18    15
#> 33: 2020-02-19    20
#> 34: 2020-02-20    23
#> 35: 2020-02-21    17
#> 36: 2020-02-22    18
#> 37: 2020-02-23    24
#> 38: 2020-02-24    18
#> 39: 2020-02-25    30
#> 40: 2020-02-26    28
#> 41: 2020-02-27    27
#> 42: 2020-02-28    38
#> 43: 2020-02-29    49
#> 44: 2020-03-01    52
#> 45: 2020-03-02    54
#> 46: 2020-03-03    56
#> 47: 2020-03-04    71
#> 48: 2020-03-05    69
#> 49: 2020-03-06    78
#> 50: 2020-03-07    78
#> 51: 2020-03-08    76
#> 52: 2020-03-09    95
#> 53: 2020-03-10    92
#> 54: 2020-03-11   108
#> 55: 2020-03-12    97
#> 56: 2020-03-13   107
#> 57: 2020-03-14   131
#>           date cases

## Mapping with a weekly reporting effect
report_weekly <- adjust_infection_to_report(
                      infections, delay_def, incubation_def,
                      reporting_effect = c(1.1, rep(1, 4), 0.95, 0.95))

#> New names:
#> * NA -> ...1
#> * NA -> ...2
#> * NA -> ...3
#> * NA -> ...4
#> * NA -> ...5
#> * ...

#> New names:
#> * NA -> ...1
#> * NA -> ...2
#> * NA -> ...3
#> * NA -> ...4
#> * NA -> ...5
#> * ...

print(report_weekly)

#>           date cases
#>  1: 2020-01-16     1
#>  2: 2020-01-17     0
#>  3: 2020-01-18     0
#>  4: 2020-01-19     0
#>  5: 2020-01-20     0
#>  6: 2020-01-21     0
#>  7: 2020-01-22     0
#>  8: 2020-01-23     0
#>  9: 2020-01-24     0
#> 10: 2020-01-25     0
#> 11: 2020-01-26     0
#> 12: 2020-01-27     0
#> 13: 2020-01-28     1
#> 14: 2020-01-29     0
#> 15: 2020-01-30     1
#> 16: 2020-01-31     0
#> 17: 2020-02-01     0
#> 18: 2020-02-02     0
#> 19: 2020-02-03     0
#> 20: 2020-02-04     2
#> 21: 2020-02-05     1
#> 22: 2020-02-06     2
#> 23: 2020-02-07     0
#> 24: 2020-02-08     3
#> 25: 2020-02-09     3
#> 26: 2020-02-10     3
#> 27: 2020-02-11     7
#> 28: 2020-02-12     6
#> 29: 2020-02-13     4
#> 30: 2020-02-14     7
#> 31: 2020-02-15     8
#> 32: 2020-02-16    22
#> 33: 2020-02-17    14
#> 34: 2020-02-18     9
#> 35: 2020-02-19    16
#> 36: 2020-02-20    23
#> 37: 2020-02-21    23
#> 38: 2020-02-22    18
#> 39: 2020-02-23    34
#> 40: 2020-02-24    22
#> 41: 2020-02-25    26
#> 42: 2020-02-26    36
#> 43: 2020-02-27    35
#> 44: 2020-02-28    38
#> 45: 2020-02-29    40
#> 46: 2020-03-01    47
#> 47: 2020-03-02    57
#> 48: 2020-03-03    48
#> 49: 2020-03-04    48
#> 50: 2020-03-05    50
#> 51: 2020-03-06    60
#> 52: 2020-03-07    57
#> 53: 2020-03-08    89
#> 54: 2020-03-09    79
#> 55: 2020-03-10    90
#> 56: 2020-03-11    97
#> 57: 2020-03-12   112
#> 58: 2020-03-13    98
#> 59: 2020-03-14   111
#>           date cases

## Map using a deterministic median shift for both delays
report_median <- adjust_infection_to_report(infections, delay_def,
                                           incubation_def, type = "median")