Fit an integer adjusted exponential or gamma distribution

Fit an integer adjusted exponential or gamma distribution

dist_fit(values = NULL, samples = NULL, dist = "exp")

Arguments

values
Numeric vector of values

samples
Numeric, number of samples to take

dist
Character string, which distribution to fit. Defaults to exponential ("exp") but gamma is also supported ("gamma").

Value

A stan fit of an interval censored distribution

Examples

dist_fit(rexp(1:100, 2), samples = 1000, dist = "exp")

#> Inference for Stan model: exp_fit.
#> 2 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=2000.
#> 
#>                mean se_mean   sd   2.5%    25%    50%    75%  97.5% n_eff Rhat
#> lambda         2.85    0.02 0.48   2.01   2.52   2.81   3.15   3.94   544 1.00
#> log_lik[1]    -0.06    0.00 0.03  -0.13  -0.07  -0.05  -0.04  -0.02   610 1.00
#> log_lik[2]    -0.29    0.00 0.05  -0.40  -0.32  -0.29  -0.26  -0.22   540 1.00
#> log_lik[3]    -0.07    0.00 0.03  -0.14  -0.08  -0.06  -0.04  -0.02   608 1.00
#> log_lik[4]    -0.04    0.00 0.02  -0.10  -0.05  -0.04  -0.03  -0.01   614 1.00
#> log_lik[5]    -0.01    0.00 0.01  -0.04  -0.02  -0.01  -0.01   0.00   627 1.00
#> log_lik[6]    -0.01    0.00 0.01  -0.03  -0.01  -0.01   0.00   0.00   630 1.00
#> log_lik[7]    -0.03    0.00 0.02  -0.07  -0.03  -0.02  -0.01  -0.01   620 1.00
#> log_lik[8]    -0.02    0.00 0.02  -0.07  -0.03  -0.02  -0.01   0.00   621 1.00
#> log_lik[9]    -0.07    0.00 0.03  -0.14  -0.08  -0.06  -0.04  -0.02   608 1.00
#> log_lik[10]   -0.03    0.00 0.02  -0.08  -0.04  -0.03  -0.02  -0.01   617 1.00
#> log_lik[11]   -0.01    0.00 0.01  -0.02  -0.01  -0.01   0.00   0.00   631 1.00
#> log_lik[12]   -0.03    0.00 0.02  -0.07  -0.03  -0.02  -0.01  -0.01   620 1.00
#> log_lik[13]   -0.02    0.00 0.01  -0.05  -0.02  -0.01  -0.01   0.00   624 1.00
#> log_lik[14]   -0.02    0.00 0.02  -0.06  -0.03  -0.02  -0.01   0.00   621 1.00
#> log_lik[15]   -0.06    0.00 0.03  -0.14  -0.08  -0.06  -0.04  -0.02   608 1.00
#> log_lik[16]   -0.86    0.01 0.14  -1.19  -0.95  -0.85  -0.77  -0.62   543 1.00
#> log_lik[17]   -1.39    0.01 0.23  -1.92  -1.54  -1.37  -1.23  -0.99   543 1.00
#> log_lik[18]   -0.05    0.00 0.03  -0.12  -0.07  -0.05  -0.03  -0.01   612 1.00
#> log_lik[19]   -2.22    0.02 0.37  -3.07  -2.46  -2.19  -1.97  -1.58   544 1.00
#> log_lik[20]   -0.06    0.00 0.03  -0.13  -0.07  -0.05  -0.04  -0.02   610 1.00
#> log_lik[21]   -0.05    0.00 0.03  -0.12  -0.07  -0.05  -0.03  -0.01   611 1.00
#> log_lik[22]   -0.06    0.00 0.03  -0.13  -0.07  -0.05  -0.04  -0.02   610 1.00
#> log_lik[23]   -0.05    0.00 0.03  -0.12  -0.07  -0.05  -0.03  -0.01   611 1.00
#> log_lik[24]   -0.04    0.00 0.02  -0.09  -0.05  -0.03  -0.02  -0.01   616 1.00
#> log_lik[25]   -0.01    0.00 0.01  -0.03  -0.01  -0.01   0.00   0.00   627 1.00
#> log_lik[26]   -0.01    0.00 0.01  -0.03  -0.01  -0.01   0.00   0.00   627 1.00
#> log_lik[27]   -0.06    0.00 0.03  -0.14  -0.08  -0.06  -0.04  -0.02   609 1.00
#> log_lik[28]   -0.02    0.00 0.02  -0.06  -0.03  -0.02  -0.01   0.00   621 1.00
#> log_lik[29]   -0.03    0.00 0.02  -0.08  -0.04  -0.03  -0.02  -0.01   617 1.00
#> log_lik[30]   -0.05    0.00 0.03  -0.12  -0.07  -0.05  -0.03  -0.01   611 1.00
#> log_lik[31]   -0.05    0.00 0.02  -0.11  -0.06  -0.04  -0.03  -0.01   613 1.00
#> log_lik[32]   -0.06    0.00 0.03  -0.14  -0.08  -0.06  -0.04  -0.02   609 1.00
#> log_lik[33]   -0.06    0.00 0.03  -0.13  -0.07  -0.05  -0.04  -0.02   610 1.00
#> log_lik[34]   -0.03    0.00 0.02  -0.07  -0.04  -0.03  -0.02  -0.01   619 1.00
#> log_lik[35]   -0.27    0.00 0.04  -0.37  -0.29  -0.26  -0.24  -0.20   539 1.00
#> log_lik[36]   -0.01    0.00 0.01  -0.02  -0.01   0.00   0.00   0.00   633 1.00
#> log_lik[37]   -1.68    0.01 0.28  -2.32  -1.86  -1.66  -1.49  -1.20   544 1.00
#> log_lik[38]   -0.02    0.00 0.02  -0.06  -0.03  -0.02  -0.01   0.00   621 1.00
#> log_lik[39]   -0.54    0.00 0.09  -0.75  -0.60  -0.53  -0.48  -0.40   542 1.00
#> log_lik[40]   -0.04    0.00 0.02  -0.09  -0.05  -0.03  -0.02  -0.01   616 1.00
#> log_lik[41]   -0.04    0.00 0.02  -0.09  -0.05  -0.03  -0.02  -0.01   616 1.00
#> log_lik[42]   -0.05    0.00 0.03  -0.11  -0.06  -0.04  -0.03  -0.01   612 1.00
#> log_lik[43]   -0.02    0.00 0.02  -0.07  -0.03  -0.02  -0.01   0.00   621 1.00
#> log_lik[44]   -0.02    0.00 0.01  -0.06  -0.03  -0.02  -0.01   0.00   622 1.00
#> log_lik[45]   -0.01    0.00 0.01  -0.04  -0.02  -0.01  -0.01   0.00   625 1.00
#> log_lik[46]   -0.06    0.00 0.03  -0.13  -0.07  -0.05  -0.04  -0.02   610 1.00
#> log_lik[47]   -0.06    0.00 0.03  -0.14  -0.08  -0.06  -0.04  -0.02   608 1.00
#> log_lik[48]   -1.30    0.01 0.21  -1.79  -1.43  -1.28  -1.15  -0.93   543 1.00
#> log_lik[49]   -0.01    0.00 0.00  -0.02  -0.01   0.00   0.00   0.00   633 1.00
#> log_lik[50]   -0.01    0.00 0.01  -0.04  -0.02  -0.01  -0.01   0.00   627 1.00
#> log_lik[51]   -0.39    0.00 0.06  -0.53  -0.43  -0.38  -0.35  -0.29   541 1.00
#> log_lik[52]   -0.05    0.00 0.03  -0.12  -0.06  -0.05  -0.03  -0.01   612 1.00
#> log_lik[53]   -0.03    0.00 0.00  -0.04  -0.03  -0.03  -0.03  -0.03   645 1.01
#> log_lik[54]   -0.05    0.00 0.02  -0.11  -0.06  -0.04  -0.03  -0.01   613 1.00
#> log_lik[55]   -0.02    0.00 0.01  -0.06  -0.03  -0.02  -0.01   0.00   623 1.00
#> log_lik[56]   -0.02    0.00 0.01  -0.04  -0.02  -0.01  -0.01   0.00   625 1.00
#> log_lik[57]   -0.03    0.00 0.02  -0.08  -0.04  -0.03  -0.02  -0.01   618 1.00
#> log_lik[58]   -0.04    0.00 0.02  -0.09  -0.05  -0.03  -0.02  -0.01   616 1.00
#> log_lik[59]   -0.01    0.00 0.01  -0.02  -0.01   0.00   0.00   0.00   632 1.00
#> log_lik[60]   -0.02    0.00 0.01  -0.05  -0.02  -0.02  -0.01   0.00   623 1.00
#> log_lik[61]   -0.02    0.00 0.02  -0.06  -0.03  -0.02  -0.01   0.00   621 1.00
#> log_lik[62]   -0.05    0.00 0.02  -0.11  -0.06  -0.04  -0.03  -0.01   613 1.00
#> log_lik[63]   -0.01    0.00 0.01  -0.03  -0.01  -0.01   0.00   0.00   627 1.00
#> log_lik[64]   -0.01    0.00 0.01  -0.03  -0.01  -0.01   0.00   0.00   627 1.00
#> log_lik[65]   -0.01    0.00 0.01  -0.03  -0.01  -0.01   0.00   0.00   629 1.00
#> log_lik[66]   -0.06    0.00 0.03  -0.13  -0.08  -0.06  -0.04  -0.02   610 1.00
#> log_lik[67]   -0.05    0.00 0.03  -0.12  -0.07  -0.05  -0.03  -0.01   611 1.00
#> log_lik[68]   -0.06    0.00 0.03  -0.13  -0.07  -0.05  -0.04  -0.02   610 1.00
#> log_lik[69]   -0.05    0.00 0.02  -0.11  -0.06  -0.04  -0.03  -0.01   613 1.00
#> log_lik[70]   -0.02    0.00 0.01  -0.05  -0.02  -0.01  -0.01   0.00   624 1.00
#> log_lik[71]   -0.05    0.00 0.03  -0.12  -0.07  -0.05  -0.03  -0.01   611 1.00
#> log_lik[72]   -0.03    0.00 0.02  -0.07  -0.04  -0.03  -0.02  -0.01   619 1.00
#> log_lik[73]   -0.04    0.00 0.02  -0.11  -0.06  -0.04  -0.03  -0.01   614 1.00
#> log_lik[74]   -0.03    0.00 0.02  -0.08  -0.04  -0.03  -0.02  -0.01   618 1.00
#> log_lik[75]   -0.05    0.00 0.03  -0.12  -0.07  -0.05  -0.03  -0.01   611 1.00
#> log_lik[76]   -0.06    0.00 0.03  -0.13  -0.07  -0.05  -0.04  -0.02   610 1.00
#> log_lik[77]   -0.01    0.00 0.01  -0.03  -0.01  -0.01   0.00   0.00   629 1.00
#> log_lik[78]   -0.02    0.00 0.01  -0.05  -0.02  -0.01  -0.01   0.00   624 1.00
#> log_lik[79]   -0.05    0.00 0.03  -0.11  -0.06  -0.04  -0.03  -0.01   612 1.00
#> log_lik[80]   -0.01    0.00 0.01  -0.04  -0.02  -0.01  -0.01   0.00   625 1.00
#> log_lik[81]   -0.03    0.00 0.02  -0.07  -0.03  -0.02  -0.01   0.00   620 1.00
#> log_lik[82]   -0.69    0.00 0.11  -0.94  -0.75  -0.67  -0.61  -0.50   542 1.00
#> log_lik[83]   -0.03    0.00 0.02  -0.07  -0.04  -0.03  -0.02  -0.01   619 1.00
#> log_lik[84]   -0.02    0.00 0.01  -0.05  -0.02  -0.01  -0.01   0.00   624 1.00
#> log_lik[85]   -0.01    0.00 0.01  -0.02  -0.01   0.00   0.00   0.00   632 1.00
#> log_lik[86]   -0.03    0.00 0.02  -0.07  -0.03  -0.02  -0.01   0.00   621 1.00
#> log_lik[87]   -0.04    0.00 0.02  -0.10  -0.05  -0.04  -0.03  -0.01   614 1.00
#> log_lik[88]   -0.02    0.00 0.01  -0.06  -0.03  -0.02  -0.01   0.00   622 1.00
#> log_lik[89]   -0.03    0.00 0.02  -0.07  -0.03  -0.02  -0.01  -0.01   620 1.00
#> log_lik[90]   -0.06    0.00 0.03  -0.12  -0.07  -0.05  -0.03  -0.01   611 1.00
#> log_lik[91]   -0.90    0.01 0.15  -1.23  -0.99  -0.88  -0.79  -0.65   543 1.00
#> log_lik[92]   -0.05    0.00 0.03  -0.12  -0.06  -0.05  -0.03  -0.01   612 1.00
#> log_lik[93]   -0.02    0.00 0.01  -0.05  -0.02  -0.01  -0.01   0.00   624 1.00
#> log_lik[94]   -0.03    0.00 0.02  -0.07  -0.04  -0.02  -0.02  -0.01   620 1.00
#> log_lik[95]   -0.01    0.00 0.01  -0.03  -0.01  -0.01   0.00   0.00   629 1.00
#> log_lik[96]   -0.04    0.00 0.02  -0.09  -0.05  -0.03  -0.02  -0.01   616 1.00
#> log_lik[97]   -0.05    0.00 0.03  -0.11  -0.06  -0.05  -0.03  -0.01   612 1.00
#> log_lik[98]   -0.04    0.00 0.02  -0.10  -0.05  -0.04  -0.02  -0.01   615 1.00
#> log_lik[99]   -0.03    0.00 0.02  -0.08  -0.04  -0.03  -0.02  -0.01   618 1.00
#> log_lik[100]  -0.06    0.00 0.03  -0.13  -0.07  -0.05  -0.04  -0.02   610 1.00
#> lp__         -12.56    0.03 0.71 -14.68 -12.68 -12.30 -12.13 -12.08   679 1.00
#> 
#> Samples were drawn using NUTS(diag_e) at Thu Jun 11 16:46:38 2020.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).