Fit an integer adjusted exponential or gamma distribution

Fit an integer adjusted exponential or gamma distribution

dist_fit(values = NULL, samples = NULL, dist = "exp")

Arguments

values
Numeric vector of values

samples
Numeric, number of samples to take

dist
Character string, which distribution to fit. Defaults to exponential ("exp") but gamma is also supported ("gamma").

Value

A stan fit of an interval censored distribution

Examples

dist_fit(rexp(1:100, 2), samples = 1000, dist = "exp")

#> Inference for Stan model: exp_fit.
#> 2 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=2000.
#> 
#>                mean se_mean   sd   2.5%    25%    50%    75%  97.5% n_eff Rhat
#> lambda         2.47    0.01 0.36   1.85   2.21   2.44   2.69   3.29   911    1
#> log_lik[1]    -0.02    0.00 0.01  -0.04  -0.02  -0.02  -0.01   0.00   935    1
#> log_lik[2]    -0.04    0.00 0.02  -0.09  -0.05  -0.04  -0.03  -0.01   934    1
#> log_lik[3]    -0.07    0.00 0.03  -0.13  -0.08  -0.06  -0.05  -0.02   933    1
#> log_lik[4]    -0.03    0.00 0.02  -0.07  -0.04  -0.03  -0.02  -0.01   934    1
#> log_lik[5]    -0.03    0.00 0.01  -0.06  -0.04  -0.03  -0.02  -0.01   935    1
#> log_lik[6]    -0.01    0.00 0.01  -0.03  -0.02  -0.01  -0.01   0.00   934    1
#> log_lik[7]    -0.07    0.00 0.03  -0.13  -0.09  -0.07  -0.05  -0.03   933    1
#> log_lik[8]    -0.01    0.00 0.01  -0.03  -0.01  -0.01  -0.01   0.00   934    1
#> log_lik[9]    -0.03    0.00 0.02  -0.07  -0.04  -0.03  -0.02  -0.01   934    1
#> log_lik[10]   -0.03    0.00 0.02  -0.07  -0.04  -0.03  -0.02  -0.01   934    1
#> log_lik[11]   -0.07    0.00 0.03  -0.13  -0.09  -0.07  -0.05  -0.02   933    1
#> log_lik[12]   -0.77    0.00 0.11  -1.02  -0.84  -0.76  -0.70  -0.60   909    1
#> log_lik[13]   -0.06    0.00 0.03  -0.12  -0.08  -0.06  -0.04  -0.02   933    1
#> log_lik[14]   -0.03    0.00 0.01  -0.06  -0.03  -0.02  -0.02  -0.01   935    1
#> log_lik[15]   -0.51    0.00 0.07  -0.67  -0.55  -0.50  -0.46  -0.40   909    1
#> log_lik[16]   -0.07    0.00 0.03  -0.14  -0.09  -0.07  -0.05  -0.03   933    1
#> log_lik[17]   -0.02    0.00 0.01  -0.05  -0.03  -0.02  -0.01   0.00   935    1
#> log_lik[18]   -0.06    0.00 0.03  -0.12  -0.08  -0.06  -0.04  -0.02   933    1
#> log_lik[19]   -1.86    0.01 0.27  -2.47  -2.02  -1.84  -1.67  -1.41   910    1
#> log_lik[20]   -0.07    0.00 0.03  -0.14  -0.09  -0.07  -0.05  -0.03   933    1
#> log_lik[21]   -0.06    0.00 0.03  -0.12  -0.08  -0.06  -0.05  -0.02   933    1
#> log_lik[22]   -0.03    0.00 0.01  -0.06  -0.03  -0.02  -0.02  -0.01   935    1
#> log_lik[23]   -0.01    0.00 0.01  -0.04  -0.02  -0.01  -0.01   0.00   935    1
#> log_lik[24]   -0.03    0.00 0.01  -0.06  -0.04  -0.03  -0.02  -0.01   935    1
#> log_lik[25]   -0.04    0.00 0.00  -0.05  -0.04  -0.04  -0.04  -0.04   935    1
#> log_lik[26]   -0.87    0.00 0.12  -1.15  -0.95  -0.86  -0.78  -0.67   910    1
#> log_lik[27]   -0.06    0.00 0.03  -0.12  -0.08  -0.06  -0.04  -0.02   933    1
#> log_lik[28]   -0.07    0.00 0.03  -0.13  -0.08  -0.06  -0.05  -0.02   933    1
#> log_lik[29]   -0.48    0.00 0.06  -0.63  -0.52  -0.47  -0.44  -0.38   908    1
#> log_lik[30]   -0.02    0.00 0.01  -0.05  -0.03  -0.02  -0.01  -0.01   935    1
#> log_lik[31]   -0.04    0.00 0.02  -0.08  -0.05  -0.03  -0.02  -0.01   934    1
#> log_lik[32]   -0.06    0.00 0.03  -0.12  -0.08  -0.06  -0.04  -0.02   933    1
#> log_lik[33]   -0.08    0.00 0.03  -0.14  -0.09  -0.07  -0.05  -0.03   933    1
#> log_lik[34]   -0.06    0.00 0.03  -0.12  -0.08  -0.06  -0.04  -0.02   933    1
#> log_lik[35]   -0.10    0.00 0.01  -0.12  -0.10  -0.09  -0.09  -0.09   894    1
#> log_lik[36]   -0.04    0.00 0.02  -0.08  -0.05  -0.04  -0.02  -0.01   934    1
#> log_lik[37]   -0.02    0.00 0.01  -0.05  -0.03  -0.02  -0.01   0.00   935    1
#> log_lik[38]   -0.15    0.00 0.02  -0.19  -0.16  -0.15  -0.14  -0.13   902    1
#> log_lik[39]   -0.07    0.00 0.03  -0.13  -0.08  -0.06  -0.05  -0.02   933    1
#> log_lik[40]   -0.08    0.00 0.03  -0.15  -0.10  -0.08  -0.06  -0.03   932    1
#> log_lik[41]   -0.07    0.00 0.03  -0.13  -0.08  -0.06  -0.05  -0.02   933    1
#> log_lik[42]   -0.01    0.00 0.01  -0.03  -0.02  -0.01  -0.01   0.00   935    1
#> log_lik[43]   -0.03    0.00 0.02  -0.07  -0.04  -0.03  -0.02  -0.01   934    1
#> log_lik[44]   -0.09    0.00 0.03  -0.16  -0.11  -0.08  -0.06  -0.03   932    1
#> log_lik[45]   -0.04    0.00 0.02  -0.08  -0.05  -0.03  -0.02  -0.01   934    1
#> log_lik[46]   -0.07    0.00 0.03  -0.13  -0.08  -0.06  -0.05  -0.02   933    1
#> log_lik[47]   -0.09    0.00 0.03  -0.16  -0.11  -0.09  -0.07  -0.04   932    1
#> log_lik[48]   -0.04    0.00 0.02  -0.08  -0.05  -0.04  -0.03  -0.01   934    1
#> log_lik[49]   -0.08    0.00 0.03  -0.15  -0.10  -0.08  -0.06  -0.03   932    1
#> log_lik[50]   -0.06    0.00 0.02  -0.11  -0.07  -0.05  -0.04  -0.02   933    1
#> log_lik[51]   -0.09    0.00 0.03  -0.16  -0.11  -0.08  -0.06  -0.03   932    1
#> log_lik[52]   -0.09    0.00 0.03  -0.16  -0.11  -0.08  -0.06  -0.03   932    1
#> log_lik[53]   -0.01    0.00 0.01  -0.03  -0.02  -0.01  -0.01   0.00   934    1
#> log_lik[54]   -0.07    0.00 0.03  -0.14  -0.09  -0.07  -0.05  -0.03   933    1
#> log_lik[55]   -0.03    0.00 0.01  -0.06  -0.04  -0.03  -0.02  -0.01   935    1
#> log_lik[56]   -0.03    0.00 0.01  -0.06  -0.03  -0.02  -0.02  -0.01   935    1
#> log_lik[57]   -0.02    0.00 0.01  -0.04  -0.02  -0.01  -0.01   0.00   935    1
#> log_lik[58]   -0.03    0.00 0.02  -0.07  -0.04  -0.03  -0.02  -0.01   934    1
#> log_lik[59]   -0.08    0.00 0.03  -0.15  -0.10  -0.08  -0.06  -0.03   932    1
#> log_lik[60]   -0.07    0.00 0.03  -0.14  -0.09  -0.07  -0.05  -0.03   933    1
#> log_lik[61]   -0.06    0.00 0.02  -0.11  -0.07  -0.05  -0.04  -0.02   933    1
#> log_lik[62]   -0.08    0.00 0.03  -0.15  -0.10  -0.08  -0.06  -0.03   932    1
#> log_lik[63]   -0.04    0.00 0.02  -0.09  -0.05  -0.04  -0.03  -0.01   934    1
#> log_lik[64]   -0.07    0.00 0.00  -0.09  -0.08  -0.07  -0.07  -0.07   889    1
#> log_lik[65]   -0.02    0.00 0.01  -0.05  -0.03  -0.02  -0.01   0.00   935    1
#> log_lik[66]   -0.07    0.00 0.03  -0.14  -0.09  -0.07  -0.05  -0.03   933    1
#> log_lik[67]   -0.06    0.00 0.02  -0.11  -0.07  -0.05  -0.04  -0.02   933    1
#> log_lik[68]   -0.06    0.00 0.03  -0.12  -0.08  -0.06  -0.04  -0.02   933    1
#> log_lik[69]   -1.74    0.01 0.25  -2.31  -1.89  -1.72  -1.56  -1.33   910    1
#> log_lik[70]   -0.04    0.00 0.02  -0.08  -0.05  -0.04  -0.03  -0.01   934    1
#> log_lik[71]   -0.08    0.00 0.03  -0.15  -0.10  -0.08  -0.06  -0.03   932    1
#> log_lik[72]   -1.66    0.01 0.24  -2.21  -1.81  -1.64  -1.49  -1.27   910    1
#> log_lik[73]   -0.03    0.00 0.01  -0.06  -0.04  -0.03  -0.02  -0.01   935    1
#> log_lik[74]   -1.92    0.01 0.28  -2.55  -2.09  -1.90  -1.72  -1.46   910    1
#> log_lik[75]   -0.04    0.00 0.02  -0.08  -0.05  -0.03  -0.02  -0.01   934    1
#> log_lik[76]   -1.20    0.01 0.17  -1.59  -1.30  -1.18  -1.08  -0.92   910    1
#> log_lik[77]   -0.02    0.00 0.01  -0.05  -0.03  -0.02  -0.01   0.00   935    1
#> log_lik[78]   -0.08    0.00 0.03  -0.15  -0.10  -0.08  -0.06  -0.03   932    1
#> log_lik[79]   -0.09    0.00 0.03  -0.16  -0.11  -0.09  -0.07  -0.04   932    1
#> log_lik[80]   -0.02    0.00 0.01  -0.05  -0.03  -0.02  -0.01   0.00   935    1
#> log_lik[81]   -0.06    0.00 0.02  -0.11  -0.07  -0.06  -0.04  -0.02   933    1
#> log_lik[82]   -0.05    0.00 0.02  -0.10  -0.06  -0.05  -0.03  -0.01   934    1
#> log_lik[83]   -0.07    0.00 0.03  -0.13  -0.09  -0.07  -0.05  -0.02   933    1
#> log_lik[84]   -0.04    0.00 0.02  -0.09  -0.06  -0.04  -0.03  -0.01   934    1
#> log_lik[85]   -0.08    0.00 0.03  -0.15  -0.10  -0.08  -0.06  -0.03   932    1
#> log_lik[86]   -0.05    0.00 0.02  -0.11  -0.07  -0.05  -0.04  -0.02   933    1
#> log_lik[87]   -0.05    0.00 0.02  -0.10  -0.06  -0.05  -0.03  -0.01   934    1
#> log_lik[88]   -0.09    0.00 0.03  -0.16  -0.11  -0.08  -0.06  -0.03   932    1
#> log_lik[89]   -0.12    0.00 0.01  -0.14  -0.12  -0.11  -0.11  -0.11   898    1
#> log_lik[90]   -0.07    0.00 0.03  -0.13  -0.08  -0.06  -0.05  -0.02   933    1
#> log_lik[91]   -0.05    0.00 0.02  -0.10  -0.07  -0.05  -0.04  -0.02   934    1
#> log_lik[92]   -0.07    0.00 0.03  -0.13  -0.08  -0.06  -0.05  -0.02   933    1
#> log_lik[93]   -0.04    0.00 0.02  -0.08  -0.05  -0.04  -0.03  -0.01   934    1
#> log_lik[94]   -0.04    0.00 0.02  -0.08  -0.05  -0.03  -0.02  -0.01   934    1
#> log_lik[95]   -1.65    0.01 0.24  -2.20  -1.80  -1.63  -1.48  -1.26   910    1
#> log_lik[96]   -0.08    0.00 0.03  -0.15  -0.10  -0.08  -0.06  -0.03   932    1
#> log_lik[97]   -0.07    0.00 0.03  -0.14  -0.09  -0.07  -0.05  -0.03   933    1
#> log_lik[98]   -0.05    0.00 0.02  -0.11  -0.07  -0.05  -0.04  -0.02   934    1
#> log_lik[99]   -0.06    0.00 0.03  -0.12  -0.08  -0.06  -0.04  -0.02   933    1
#> log_lik[100]  -1.34    0.01 0.19  -1.78  -1.46  -1.32  -1.20  -1.02   910    1
#> lp__         -17.96    0.02 0.65 -19.83 -18.08 -17.72 -17.56 -17.51   908    1
#> 
#> Samples were drawn using NUTS(diag_e) at Thu May 28 13:23:11 2020.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).