Estimate time-varying measures and forecast

Estimate time-varying measures and forecast

epi_measures_pipeline(
  nowcast = NULL,
  generation_times = NULL,
  min_est_date = NULL,
  gt_samples = 1,
  rt_samples = 5,
  rt_windows = 7,
  rate_window = 7,
  rt_prior = NULL,
  forecast_model = NULL,
  horizon = 0,
  verbose = TRUE
)

Arguments

nowcast
A nowcast as produced by nowcast_pipeline

generation_times
A matrix with columns representing samples and rows representing the probability of the generation timebeing on that day.

min_est_date
Date to begin estimation.

gt_samples
Numeric, the number of samples to take from the generaiton times supplied

rt_samples
Numeric, the number of samples to take from the estimated R distribution for each time point.

rt_windows
Numeric vector, windows over which to estimate time-varying R. The best performing window will be selected per serial interval sample by default (based on which window best forecasts current cases).

rate_window
Numeric, the window to use to estimate the rate of spread.

rt_prior
A list defining the reproduction number prior containing the mean (mean_prior) and standard deviation (std_prior)

forecast_model
An uninitialised bsts model passed to EpiSoon::forecast_rt to be used for forecasting future Rt values. An example of the required structure is: function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}.

horizon
Numeric, defaults to 0. The horizon over which to forecast Rts and cases.

verbose
Logical, defaults to TRUE. Should progress messages be shown.

Value

A list of data frames containing reproduction number estimates, case forecasts, rate of growth estimates both summarised and raw.

Examples

## Construct example distributions
## reporting delay dist
delay_dist <- EpiNow::lognorm_dist_def(mean = 3,
                                      mean_sd = 1,
                                      sd = 3,
                                      sd_sd = 1,
                                      max_value = 30,
                                      samples = 1)

## incubation delay dist
incubation_dist <- delay_dist

## Uses example case vector from EpiSoon
cases <- data.table::setDT(EpiSoon::example_obs_cases)
cases <- cases[, `:=`(confirm = as.integer(cases), import_status = "local")]

## Basic nowcast
nowcast <- nowcast_pipeline(reported_cases = cases,
                            target_date = max(cases$date),
                            delay_defs = delay_dist,
                            incubation_defs = incubation_dist)

#> New names:
#> * NA -> ...1
#> * NA -> ...2
#> * NA -> ...3
#> * NA -> ...4
#> * NA -> ...5
#> * ...

## Estimate parameters                           
estimates <- epi_measures_pipeline(nowcast[type %in% "infection_upscaled"],
                       generation_times = EpiNow::covid_generation_times,
                       rt_prior = list(mean_prior = 2.6, std_prior = 2))

#> Estimate time-varying R0

#> Summarising time-varying R0

#> Summarising forecast cases

#> Estimate time-varying rate of growth

#> 
#> Attaching package: ‘purrr’

#> The following object is masked from ‘package:data.table’:
#> 
#>     transpose

estimates

#> $R0
#>                   type       date rt_type    bottom      top     lower
#>  1: infection_upscaled 2020-01-16 nowcast 0.7521816 1.364256 1.2800466
#>  2: infection_upscaled 2020-01-17 nowcast 0.2417629 3.346998 0.2417629
#>  3: infection_upscaled 2020-01-18 nowcast 0.5958374 1.911790 0.5958374
#>  4: infection_upscaled 2020-01-19 nowcast 0.4228348 1.810311 0.4228348
#>  5: infection_upscaled 2020-01-20 nowcast 0.7504479 2.049361 1.7337573
#>  6: infection_upscaled 2020-01-21 nowcast 0.7808342 2.642001 0.7808342
#>  7: infection_upscaled 2020-01-22 nowcast 1.0416352 2.135777 1.7548286
#>  8: infection_upscaled 2020-01-23 nowcast 1.3742518 5.231693 3.4097832
#>  9: infection_upscaled 2020-01-24 nowcast 1.7843516 3.651134 2.5016697
#> 10: infection_upscaled 2020-01-25 nowcast 1.5180530 4.802774 1.5180530
#> 11: infection_upscaled 2020-01-26 nowcast 1.2440076 3.692613 1.2440076
#> 12: infection_upscaled 2020-01-27 nowcast 1.7624932 6.419534 2.2394177
#> 13: infection_upscaled 2020-01-28 nowcast 1.0576397 3.112196 1.0576397
#> 14: infection_upscaled 2020-01-29 nowcast 1.0421863 2.392305 1.4263140
#> 15: infection_upscaled 2020-01-30 nowcast 2.2089893 3.286969 2.2089893
#> 16: infection_upscaled 2020-01-31 nowcast 2.2018759 3.561709 2.6898910
#> 17: infection_upscaled 2020-02-01 nowcast 1.8838460 3.727824 2.1589266
#> 18: infection_upscaled 2020-02-02 nowcast 2.2371687 3.107562 2.2371687
#> 19: infection_upscaled 2020-02-03 nowcast 2.0348471 2.846491 2.4431679
#> 20: infection_upscaled 2020-02-04 nowcast 1.5447830 2.209270 1.5447830
#> 21: infection_upscaled 2020-02-05 nowcast 1.4228438 2.174256 1.4228438
#> 22: infection_upscaled 2020-02-06 nowcast 1.5728248 2.187454 1.9964160
#> 23: infection_upscaled 2020-02-07 nowcast 0.8979752 2.140033 0.8979752
#> 24: infection_upscaled 2020-02-08 nowcast 1.0834113 1.875654 1.0834113
#> 25: infection_upscaled 2020-02-09 nowcast 1.2007590 1.595897 1.2295261
#> 26: infection_upscaled 2020-02-10 nowcast 1.2936845 2.496972 1.7302055
#> 27: infection_upscaled 2020-02-11 nowcast 1.4958939 1.896916 1.7416821
#> 28: infection_upscaled 2020-02-12 nowcast 1.5233666 2.080063 1.5233666
#> 29: infection_upscaled 2020-02-13 nowcast 1.4160374 2.019370 1.4160374
#> 30: infection_upscaled 2020-02-14 nowcast 1.5049646 1.787779 1.7836744
#> 31: infection_upscaled 2020-02-15 nowcast 1.7783094 2.539088 1.7783094
#> 32: infection_upscaled 2020-02-16 nowcast 1.6223779 2.102501 1.8759813
#> 33: infection_upscaled 2020-02-17 nowcast 1.4343576 1.906383 1.6607371
#> 34: infection_upscaled 2020-02-18 nowcast 1.3755471 2.034707 1.3755471
#> 35: infection_upscaled 2020-02-19 nowcast 1.6388382 1.852887 1.6388382
#> 36: infection_upscaled 2020-02-20 nowcast 1.4396389 1.663583 1.4396389
#> 37: infection_upscaled 2020-02-21 nowcast 1.4031076 1.909551 1.6875830
#> 38: infection_upscaled 2020-02-22 nowcast 1.2987883 1.758600 1.2987883
#> 39: infection_upscaled 2020-02-23 nowcast 1.4371110 1.759878 1.4371110
#> 40: infection_upscaled 2020-02-24 nowcast 1.5469449 1.835106 1.6681793
#> 41: infection_upscaled 2020-02-25 nowcast 1.5989932 1.730625 1.7127783
#> 42: infection_upscaled 2020-02-26 nowcast 1.4633025 1.864702 1.7386230
#> 43: infection_upscaled 2020-02-27 nowcast 1.4082753 1.672026 1.4820034
#> 44: infection_upscaled 2020-02-28 nowcast 1.5017332 1.663685 1.5889388
#> 45: infection_upscaled 2020-02-29 nowcast 1.5024534 1.653237 1.5024534
#> 46: infection_upscaled 2020-03-01 nowcast 1.4158169 1.620193 1.4158169
#> 47: infection_upscaled 2020-03-02 nowcast 1.3186499 1.451391 1.4051626
#> 48: infection_upscaled 2020-03-03 nowcast 1.3633952 1.582845 1.4078683
#> 49: infection_upscaled 2020-03-04 nowcast 1.1964025 1.455513 1.1964025
#> 50: infection_upscaled 2020-03-05 nowcast 1.2375223 1.378126 1.2375223
#> 51: infection_upscaled 2020-03-06 nowcast 1.2829169 1.457405 1.2829169
#> 52: infection_upscaled 2020-03-07 nowcast 1.1925116 1.324041 1.2867411
#> 53: infection_upscaled 2020-03-08 nowcast 1.2217592 1.347332 1.3089322
#> 54: infection_upscaled 2020-03-09 nowcast 1.2014049 1.333665 1.2179085
#> 55: infection_upscaled 2020-03-10 nowcast 1.2939045 1.408541 1.2939045
#> 56: infection_upscaled 2020-03-11 nowcast 1.2368693 1.350215 1.3063981
#> 57: infection_upscaled 2020-03-12 nowcast 1.2181386 1.303926 1.2616769
#> 58: infection_upscaled 2020-03-13 nowcast 1.0947383 1.330614 1.2678473
#> 59: infection_upscaled 2020-03-14 nowcast 1.1667134 1.301867 1.2520468
#>                   type       date rt_type    bottom      top     lower
#>         upper    median     mean        std prob_control mean_window sd_window
#>  1: 1.3642556 1.2800466 1.117940 0.29997565          0.4           7         0
#>  2: 0.3908169 0.3908169 1.261617 1.39669260          0.6           7         0
#>  3: 0.9172158 0.9172158 1.107769 0.54445871          0.6           7         0
#>  4: 0.6869218 0.6869218 1.019976 0.59869592          0.6           7         0
#>  5: 2.0493613 1.7337573 1.456731 0.56181358          0.4           7         0
#>  6: 1.1794785 1.1794785 1.495080 0.75151227          0.2           7         0
#>  7: 2.1357768 1.7548286 1.555860 0.48089254          0.0           7         0
#>  8: 5.2316929 3.4097832 3.328765 1.85635322          0.0           7         0
#>  9: 2.9754334 2.8232451 2.747167 0.68233649          0.0           7         0
#> 10: 2.0265430 2.0265430 2.923609 1.49492036          0.0           7         0
#> 11: 1.9825859 1.9825859 2.319847 0.92605740          0.0           7         0
#> 12: 3.7474815 3.5013427 3.534054 1.81570402          0.0           7         0
#> 13: 1.6975498 1.6975498 1.954142 0.89796360          0.0           7         0
#> 14: 1.9278167 1.5487922 1.667483 0.51390328          0.0           7         0
#> 15: 2.3833705 2.3833705 2.579432 0.44681004          0.0           7         0
#> 16: 2.8585633 2.8253323 2.827474 0.48743585          0.0           7         0
#> 17: 2.5241479 2.2656030 2.512069 0.71734320          0.0           7         0
#> 18: 2.4867479 2.4867479 2.575729 0.34239516          0.0           7         0
#> 19: 2.8464909 2.4431679 2.507917 0.34201472          0.0           7         0
#> 20: 1.7467878 1.7467878 1.841653 0.30865124          0.0           7         0
#> 21: 1.5736066 1.5736066 1.691940 0.30706871          0.0           7         0
#> 22: 2.1874538 1.9964160 1.909858 0.25015779          0.0           7         0
#> 23: 1.1663896 1.1663896 1.339736 0.50513355          0.4           7         0
#> 24: 1.4432436 1.4432436 1.406652 0.32112197          0.0           7         0
#> 25: 1.3549052 1.3463587 1.345489 0.15587461          0.0           7         0
#> 26: 2.0401554 1.9250248 1.897208 0.43961781          0.0           7         0
#> 27: 1.8969159 1.7416821 1.695377 0.15712117          0.0           7         0
#> 28: 1.7419048 1.7419048 1.810778 0.24585883          0.0           7         0
#> 29: 1.6562314 1.6562314 1.732690 0.24345547          0.0           7         0
#> 30: 1.7877785 1.7836744 1.719847 0.12178581          0.0           7         0
#> 31: 1.9487768 1.9487768 2.049612 0.29721581          0.0           7         0
#> 32: 2.1025012 1.8759813 1.879868 0.19905338          0.0           7         0
#> 33: 1.8447011 1.6857814 1.706392 0.18414502          0.0           7         0
#> 34: 1.6262123 1.6262123 1.620770 0.27064241          0.0           7         0
#> 35: 1.7216260 1.7216260 1.746598 0.08538349          0.0           7         0
#> 36: 1.5416773 1.5416773 1.548079 0.10708589          0.0           7         0
#> 37: 1.9095507 1.6875830 1.673519 0.21038922          0.0           7         0
#> 38: 1.3490429 1.3490429 1.468251 0.20680136          0.0           7         0
#> 39: 1.5453027 1.5453027 1.577851 0.14082013          0.0           7         0
#> 40: 1.7569263 1.6835856 1.698149 0.10745359          0.0           7         0
#> 41: 1.7306248 1.7127783 1.688715 0.05236189          0.0           7         0
#> 42: 1.8647021 1.7386230 1.677765 0.15622396          0.0           7         0
#> 43: 1.5134015 1.4995235 1.515046 0.09666717          0.0           7         0
#> 44: 1.6636849 1.5889388 1.599079 0.06766286          0.0           7         0
#> 45: 1.5335985 1.5335985 1.551319 0.06116721          0.0           7         0
#> 46: 1.5172607 1.5172607 1.507005 0.08512891          0.0           7         0
#> 47: 1.4513911 1.4051626 1.390023 0.05080332          0.0           7         0
#> 48: 1.4474163 1.4392849 1.448162 0.08220077          0.0           7         0
#> 49: 1.2150421 1.2150421 1.290487 0.12008185          0.0           7         0
#> 50: 1.2752276 1.2752276 1.302370 0.07053735          0.0           7         0
#> 51: 1.3007922 1.3007922 1.341193 0.07341552          0.0           7         0
#> 52: 1.3240412 1.2867411 1.265323 0.05445798          0.0           7         0
#> 53: 1.3473319 1.3089322 1.301369 0.05123233          0.0           7         0
#> 54: 1.2413015 1.2367133 1.246199 0.05140790          0.0           7         0
#> 55: 1.3338033 1.3338033 1.345539 0.04868172          0.0           7         0
#> 56: 1.3502145 1.3063981 1.301739 0.04456894          0.0           7         0
#> 57: 1.3039259 1.2616769 1.259696 0.03569978          0.0           7         0
#> 58: 1.3306136 1.2678473 1.232812 0.10316569          0.0           7         0
#> 59: 1.3018674 1.2520468 1.233392 0.05512822          0.0           7         0
#>         upper    median     mean        std prob_control mean_window sd_window
#>     mean_crps sd_crps R0_range
#>  1:      0.04       0   <list>
#>  2:      0.04       0   <list>
#>  3:      0.04       0   <list>
#>  4:      0.00       0   <list>
#>  5:      0.00       0   <list>
#>  6:      1.00       0   <list>
#>  7:      0.04       0   <list>
#>  8:      1.36       0   <list>
#>  9:      0.36       0   <list>
#> 10:      0.76       0   <list>
#> 11:      0.80       0   <list>
#> 12:      1.00       0   <list>
#> 13:      0.68       0   <list>
#> 14:      0.12       0   <list>
#> 15:      2.08       0   <list>
#> 16:      1.52       0   <list>
#> 17:      1.16       0   <list>
#> 18:      1.12       0   <list>
#> 19:      3.16       0   <list>
#> 20:      1.00       0   <list>
#> 21:      3.60       0   <list>
#> 22:      1.84       0   <list>
#> 23:      1.08       0   <list>
#> 24:      2.56       0   <list>
#> 25:      1.48       0   <list>
#> 26:      2.20       0   <list>
#> 27:      0.24       0   <list>
#> 28:      3.16       0   <list>
#> 29:      2.60       0   <list>
#> 30:      0.36       0   <list>
#> 31:      0.52       0   <list>
#> 32:      2.44       0   <list>
#> 33:      0.96       0   <list>
#> 34:      3.04       0   <list>
#> 35:      2.68       0   <list>
#> 36:      1.32       0   <list>
#> 37:      2.32       0   <list>
#> 38:      2.00       0   <list>
#> 39:      5.52       0   <list>
#> 40:      1.80       0   <list>
#> 41:      4.28       0   <list>
#> 42:      2.36       0   <list>
#> 43:      5.32       0   <list>
#> 44:      2.36       0   <list>
#> 45:      9.24       0   <list>
#> 46:      7.80       0   <list>
#> 47:      4.80       0   <list>
#> 48:      2.72       0   <list>
#> 49:      3.92       0   <list>
#> 50:      4.08       0   <list>
#> 51:      3.24       0   <list>
#> 52:      4.24       0   <list>
#> 53:      8.40       0   <list>
#> 54:      5.96       0   <list>
#> 55:      2.20       0   <list>
#> 56:      6.00       0   <list>
#> 57:      3.00       0   <list>
#> 58:     10.92       0   <list>
#> 59:     13.88       0   <list>
#>     mean_crps sd_crps R0_range
#> 
#> $rate_of_spread
#>                  type overall_little_r time_varying_r
#> 1: infection_upscaled           <list>   <data.table>
#> 
#> $raw_R0
#>            date         R sample  crps window rt_type               type
#>   1: 2020-01-16 0.7521816      1  0.04      7 nowcast infection_upscaled
#>   2: 2020-01-16 0.8336432      1  0.04      7 nowcast infection_upscaled
#>   3: 2020-01-16 1.2800466      1  0.04      7 nowcast infection_upscaled
#>   4: 2020-01-16 1.3595756      1  0.04      7 nowcast infection_upscaled
#>   5: 2020-01-16 1.3642556      1  0.04      7 nowcast infection_upscaled
#>  ---                                                                    
#> 291: 2020-03-14 1.1667134      1 13.88      7 nowcast infection_upscaled
#> 292: 2020-03-14 1.1880626      1 13.88      7 nowcast infection_upscaled
#> 293: 2020-03-14 1.2520468      1 13.88      7 nowcast infection_upscaled
#> 294: 2020-03-14 1.2582678      1 13.88      7 nowcast infection_upscaled
#> 295: 2020-03-14 1.3018674      1 13.88      7 nowcast infection_upscaled
#>