Estimate time-varying measures and forecast

Estimate time-varying measures and forecast

epi_measures_pipeline(
  nowcast = NULL,
  generation_times = NULL,
  min_est_date = NULL,
  gt_samples = 1,
  rt_samples = 5,
  rt_windows = 7,
  rate_window = 7,
  rt_prior = NULL,
  forecast_model = NULL,
  horizon = 0,
  verbose = TRUE
)

Arguments

nowcast
A nowcast as produced by nowcast_pipeline

generation_times
A matrix with columns representing samples and rows representing the probability of the generation timebeing on that day.

min_est_date
Date to begin estimation.

gt_samples
Numeric, the number of samples to take from the generaiton times supplied

rt_samples
Numeric, the number of samples to take from the estimated R distribution for each time point.

rt_windows
Numeric vector, windows over which to estimate time-varying R. The best performing window will be selected per serial interval sample by default (based on which window best forecasts current cases).

rate_window
Numeric, the window to use to estimate the rate of spread.

rt_prior
A list defining the reproduction number prior containing the mean (mean_prior) and standard deviation (std_prior)

forecast_model
An uninitialised bsts model passed to EpiSoon::forecast_rt to be used for forecasting future Rt values. An example of the required structure is: function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}.

horizon
Numeric, defaults to 0. The horizon over which to forecast Rts and cases.

verbose
Logical, defaults to TRUE. Should progress messages be shown.

Value

A list of data frames containing reproduction number estimates, case forecasts, rate of growth estimates both summarised and raw.

Examples

## Construct example distributions
## reporting delay dist
delay_dist <- EpiNow::lognorm_dist_def(mean = 3,
                                      mean_sd = 1,
                                      sd = 3,
                                      sd_sd = 1,
                                      max_value = 30,
                                      samples = 1)

## incubation delay dist
incubation_dist <- delay_dist

## Uses example case vector from EpiSoon
cases <- data.table::setDT(EpiSoon::example_obs_cases)
cases <- cases[, `:=`(confirm = as.integer(cases), import_status = "local")]

## Basic nowcast
nowcast <- nowcast_pipeline(reported_cases = cases,
                            target_date = max(cases$date),
                            delay_defs = delay_dist,
                            incubation_defs = incubation_dist)

## Estimate parameters                           
estimates <- epi_measures_pipeline(nowcast[type %in% "infection_upscaled"],
                       generation_times = EpiNow::covid_generation_times,
                       rt_prior = list(mean_prior = 2.6, std_prior = 2))

#> Estimate time-varying R0

#> Summarising time-varying R0

#> Summarising forecast cases

#> Estimate time-varying rate of growth

#> 
#> Attaching package: ‘purrr’

#> The following object is masked from ‘package:data.table’:
#> 
#>     transpose

estimates

#> $R0
#>                   type       date rt_type    bottom      top     lower
#>  1: infection_upscaled 2020-01-05 nowcast 0.5111961 2.201969 1.5000929
#>  2: infection_upscaled 2020-01-06 nowcast 0.6577401 1.364343 0.6577401
#>  3: infection_upscaled 2020-01-07 nowcast 1.0027356 3.035099 1.0027356
#>  4: infection_upscaled 2020-01-08 nowcast 0.1880813 3.935860 0.5383833
#>  5: infection_upscaled 2020-01-09 nowcast 0.1680303 5.198515 0.5618973
#>  6: infection_upscaled 2020-01-10 nowcast 0.3814451 1.444274 0.8606341
#>  7: infection_upscaled 2020-01-11 nowcast 0.6725188 6.267806 1.1647662
#>  8: infection_upscaled 2020-01-12 nowcast 3.4892169 6.147526 3.4892169
#>  9: infection_upscaled 2020-01-13 nowcast 1.3272305 3.198537 2.1497246
#> 10: infection_upscaled 2020-01-14 nowcast 0.9202680 5.458532 0.9202680
#> 11: infection_upscaled 2020-01-15 nowcast 1.3884427 3.793225 1.3884427
#> 12: infection_upscaled 2020-01-16 nowcast 1.9423941 4.108476 1.9423941
#> 13: infection_upscaled 2020-01-17 nowcast 2.0345955 5.658004 3.6616988
#> 14: infection_upscaled 2020-01-18 nowcast 1.8957272 5.925870 2.3948196
#> 15: infection_upscaled 2020-01-19 nowcast 1.3264484 3.133510 2.3271005
#> 16: infection_upscaled 2020-01-20 nowcast 1.9501930 3.639519 1.9501930
#> 17: infection_upscaled 2020-01-21 nowcast 1.2188070 5.047450 2.4532205
#> 18: infection_upscaled 2020-01-22 nowcast 1.6677863 2.490588 1.6677863
#> 19: infection_upscaled 2020-01-23 nowcast 0.9597757 2.085256 0.9597757
#> 20: infection_upscaled 2020-01-24 nowcast 0.7492985 4.879686 1.2521227
#> 21: infection_upscaled 2020-01-25 nowcast 1.8265471 3.066970 1.8265471
#> 22: infection_upscaled 2020-01-26 nowcast 1.5428275 2.361699 2.2116629
#> 23: infection_upscaled 2020-01-27 nowcast 1.5782056 2.615520 1.8066139
#> 24: infection_upscaled 2020-01-28 nowcast 1.3018317 2.531031 1.3018317
#> 25: infection_upscaled 2020-01-29 nowcast 1.6203088 2.617766 2.1096598
#> 26: infection_upscaled 2020-01-30 nowcast 1.4479702 2.085004 1.7707279
#> 27: infection_upscaled 2020-01-31 nowcast 1.5341370 2.416721 1.5341370
#> 28: infection_upscaled 2020-02-01 nowcast 1.5722756 2.192669 2.0974356
#> 29: infection_upscaled 2020-02-02 nowcast 1.5621385 2.180298 1.5621385
#> 30: infection_upscaled 2020-02-03 nowcast 1.6986157 2.021878 1.9422502
#> 31: infection_upscaled 2020-02-04 nowcast 1.6697451 2.109402 1.6697451
#> 32: infection_upscaled 2020-02-05 nowcast 1.7988248 2.220235 1.7988248
#> 33: infection_upscaled 2020-02-06 nowcast 1.7525906 2.069924 1.9665225
#> 34: infection_upscaled 2020-02-07 nowcast 1.9103329 2.048303 2.0112593
#> 35: infection_upscaled 2020-02-08 nowcast 1.3788883 1.869097 1.6055088
#> 36: infection_upscaled 2020-02-09 nowcast 1.4143279 1.789806 1.6289789
#> 37: infection_upscaled 2020-02-10 nowcast 1.6099592 1.713450 1.6497655
#> 38: infection_upscaled 2020-02-11 nowcast 1.4504784 1.748523 1.6714147
#> 39: infection_upscaled 2020-02-12 nowcast 1.3205106 1.722885 1.3205106
#> 40: infection_upscaled 2020-02-13 nowcast 1.3083300 1.646456 1.3861627
#> 41: infection_upscaled 2020-02-14 nowcast 1.2521779 1.474790 1.4369060
#> 42: infection_upscaled 2020-02-15 nowcast 1.4172575 1.761772 1.4172575
#> 43: infection_upscaled 2020-02-16 nowcast 1.4574103 1.579103 1.5590102
#> 44: infection_upscaled 2020-02-17 nowcast 1.4455481 1.566715 1.4455481
#> 45: infection_upscaled 2020-02-18 nowcast 1.4080597 1.618795 1.4512102
#> 46: infection_upscaled 2020-02-19 nowcast 1.3099798 1.665740 1.4566413
#> 47: infection_upscaled 2020-02-20 nowcast 1.4489470 1.674519 1.5864237
#> 48: infection_upscaled 2020-02-21 nowcast 1.5051536 1.615402 1.5051536
#> 49: infection_upscaled 2020-02-22 nowcast 1.4823703 1.659504 1.4823703
#> 50: infection_upscaled 2020-02-23 nowcast 1.5098324 1.573279 1.5098324
#> 51: infection_upscaled 2020-02-24 nowcast 1.4522608 1.560600 1.4522608
#> 52: infection_upscaled 2020-02-25 nowcast 1.4743643 1.639894 1.4845943
#> 53: infection_upscaled 2020-02-26 nowcast 1.4102869 1.502844 1.4753851
#> 54: infection_upscaled 2020-02-27 nowcast 1.4418358 1.513674 1.4418358
#> 55: infection_upscaled 2020-02-28 nowcast 1.2978978 1.425666 1.3798666
#> 56: infection_upscaled 2020-02-29 nowcast 1.3235146 1.387010 1.3235146
#> 57: infection_upscaled 2020-03-01 nowcast 1.2377314 1.384125 1.2377314
#> 58: infection_upscaled 2020-03-02 nowcast 1.2053671 1.295458 1.2549982
#> 59: infection_upscaled 2020-03-03 nowcast 1.2237741 1.307567 1.2415487
#> 60: infection_upscaled 2020-03-04 nowcast 1.1555106 1.271321 1.2410780
#> 61: infection_upscaled 2020-03-05 nowcast 1.1801930 1.222491 1.2076734
#> 62: infection_upscaled 2020-03-06 nowcast 1.1688323 1.296863 1.2203873
#> 63: infection_upscaled 2020-03-07 nowcast 1.1950243 1.264217 1.2173680
#> 64: infection_upscaled 2020-03-08 nowcast 1.1850821 1.259796 1.2263166
#> 65: infection_upscaled 2020-03-09 nowcast 1.2090510 1.293458 1.2482502
#> 66: infection_upscaled 2020-03-10 nowcast 1.2392664 1.345212 1.2392664
#> 67: infection_upscaled 2020-03-11 nowcast 1.1951453 1.250850 1.1983335
#> 68: infection_upscaled 2020-03-12 nowcast 1.2079866 1.278209 1.2669756
#> 69: infection_upscaled 2020-03-13 nowcast 1.2085870 1.269137 1.2085870
#> 70: infection_upscaled 2020-03-14 nowcast 1.1638980 1.306561 1.1638980
#>                   type       date rt_type    bottom      top     lower
#>         upper    median      mean        std prob_control mean_window sd_window
#>  1: 2.2019690 1.5000929 1.4555807 0.74685048          0.4           7         0
#>  2: 0.7844008 0.7844008 0.9059955 0.29478817          0.6           7         0
#>  3: 1.4433535 1.4433535 1.8021461 0.86754321          0.0           7         0
#>  4: 1.4500630 1.4188437 1.5062463 1.46527623          0.4           7         0
#>  5: 0.8288605 0.6245422 1.4763691 2.09447228          0.8           7         0
#>  6: 1.1454329 1.0488882 0.9761349 0.39373395          0.4           7         0
#>  7: 2.0565511 1.7920589 2.3907402 2.23366498          0.2           7         0
#>  8: 3.8402010 3.8402010 4.5579928 1.24048453          0.0           7         0
#>  9: 2.5646042 2.2082815 2.2896756 0.68083159          0.0           7         0
#> 10: 2.5403429 2.5403429 2.7793843 1.86803227          0.2           7         0
#> 11: 2.4137886 2.4137886 2.5703716 1.02889314          0.0           7         0
#> 12: 2.5662349 2.5662349 2.8958691 0.89313337          0.0           7         0
#> 13: 4.1775946 3.9013902 3.8866567 1.29453375          0.0           7         0
#> 14: 3.9432525 3.7615844 3.5842507 1.57395744          0.0           7         0
#> 15: 3.1335100 2.3271005 2.1023954 0.76308413          0.0           7         0
#> 16: 2.1378595 2.1378595 2.5588041 0.75115698          0.0           7         0
#> 17: 2.8136253 2.5664790 2.8199163 1.38962148          0.0           7         0
#> 18: 1.8893590 1.8893590 2.0268147 0.34988661          0.0           7         0
#> 19: 0.9927661 0.9927661 1.3274095 0.51245185          0.6           7         0
#> 20: 2.2958453 2.0198131 2.2393531 1.59825169          0.2           7         0
#> 21: 2.2657434 2.2657434 2.4870410 0.54662787          0.0           7         0
#> 22: 2.3616987 2.2116629 2.0539792 0.35650315          0.0           7         0
#> 23: 2.0664064 1.9766122 2.0086716 0.38689604          0.0           7         0
#> 24: 1.6351275 1.6351275 1.8493776 0.58803845          0.0           7         0
#> 25: 2.1796970 2.1188133 2.1292491 0.35380326          0.0           7         0
#> 26: 2.0850040 1.7707279 1.7661273 0.25857677          0.0           7         0
#> 27: 1.9377214 1.9377214 1.9968564 0.40335482          0.0           7         0
#> 28: 2.1926690 2.0974356 1.9302718 0.31766454          0.0           7         0
#> 29: 1.6980466 1.6980466 1.7511859 0.25034945          0.0           7         0
#> 30: 2.0218778 1.9422502 1.8664137 0.15373483          0.0           7         0
#> 31: 1.7498215 1.7498215 1.8224368 0.17175903          0.0           7         0
#> 32: 1.9296677 1.9296677 1.9752773 0.18904885          0.0           7         0
#> 33: 2.0699242 1.9665225 1.9396409 0.13172359          0.0           7         0
#> 34: 2.0483034 2.0112593 1.9818629 0.06323666          0.0           7         0
#> 35: 1.7093238 1.6997522 1.6525140 0.17989410          0.0           7         0
#> 36: 1.6822266 1.6627816 1.6356241 0.13756579          0.0           7         0
#> 37: 1.6999769 1.6609924 1.6668289 0.04132716          0.0           7         0
#> 38: 1.7485227 1.6714147 1.6287477 0.12684473          0.0           7         0
#> 39: 1.4997158 1.4997158 1.5007926 0.16023483          0.0           7         0
#> 40: 1.5134922 1.4995313 1.4707945 0.12948282          0.0           7         0
#> 41: 1.4747905 1.4369060 1.3880117 0.09527832          0.0           7         0
#> 42: 1.5097785 1.5097785 1.5623312 0.13886834          0.0           7         0
#> 43: 1.5697035 1.5673739 1.5465203 0.05032541          0.0           7         0
#> 44: 1.5017656 1.5017656 1.5122604 0.05266883          0.0           7         0
#> 45: 1.5291263 1.5005057 1.5015395 0.08028473          0.0           7         0
#> 46: 1.5382683 1.5347194 1.5010698 0.13050058          0.0           7         0
#> 47: 1.6745190 1.5864237 1.5645187 0.10640589          0.0           7         0
#> 48: 1.5597781 1.5597781 1.5631751 0.04756887          0.0           7         0
#> 49: 1.5532917 1.5532917 1.5679873 0.06884963          0.0           7         0
#> 50: 1.5255352 1.5255352 1.5327717 0.02533919          0.0           7         0
#> 51: 1.4895748 1.4895748 1.5067216 0.04419347          0.0           7         0
#> 52: 1.4970361 1.4908790 1.5173535 0.06901270          0.0           7         0
#> 53: 1.4920556 1.4812312 1.4723607 0.03625649          0.0           7         0
#> 54: 1.4718245 1.4718245 1.4750378 0.03279197          0.0           7         0
#> 55: 1.4256661 1.3798666 1.3748613 0.05037740          0.0           7         0
#> 56: 1.3509092 1.3509092 1.3582205 0.02645464          0.0           7         0
#> 57: 1.2748758 1.2748758 1.3009940 0.06140939          0.0           7         0
#> 58: 1.2954578 1.2549982 1.2531528 0.03723620          0.0           7         0
#> 59: 1.2613683 1.2561363 1.2580788 0.03127975          0.0           7         0
#> 60: 1.2713213 1.2410780 1.2230264 0.04648961          0.0           7         0
#> 61: 1.2224908 1.2076734 1.2015412 0.01950696          0.0           7         0
#> 62: 1.2503320 1.2229332 1.2318696 0.04677182          0.0           7         0
#> 63: 1.2412573 1.2362271 1.2308188 0.02605684          0.0           7         0
#> 64: 1.2597956 1.2263166 1.2229161 0.03048075          0.0           7         0
#> 65: 1.2837935 1.2516443 1.2572393 0.03335837          0.0           7         0
#> 66: 1.2735953 1.2735953 1.2864616 0.04412392          0.0           7         0
#> 67: 1.2110469 1.2087331 1.2128218 0.02229573          0.0           7         0
#> 68: 1.2782088 1.2669756 1.2517476 0.02942054          0.0           7         0
#> 69: 1.2320000 1.2320000 1.2370390 0.02445141          0.0           7         0
#> 70: 1.2003160 1.2003160 1.2071997 0.05851703          0.0           7         0
#>         upper    median      mean        std prob_control mean_window sd_window
#>     mean_crps sd_crps R0_range
#>  1:      0.04       0   <list>
#>  2:      0.00       0   <list>
#>  3:      0.00       0   <list>
#>  4:      0.00       0   <list>
#>  5:      0.00       0   <list>
#>  6:      0.00       0   <list>
#>  7:      0.00       0   <list>
#>  8:      1.00       0   <list>
#>  9:      0.00       0   <list>
#> 10:      0.04       0   <list>
#> 11:      0.04       0   <list>
#> 12:      0.64       0   <list>
#> 13:      0.84       0   <list>
#> 14:      0.20       0   <list>
#> 15:      0.64       0   <list>
#> 16:      0.44       0   <list>
#> 17:      1.32       0   <list>
#> 18:      0.36       0   <list>
#> 19:      0.80       0   <list>
#> 20:      1.08       0   <list>
#> 21:      2.08       0   <list>
#> 22:      0.52       0   <list>
#> 23:      0.88       0   <list>
#> 24:      0.64       0   <list>
#> 25:      0.76       0   <list>
#> 26:      1.56       0   <list>
#> 27:      1.72       0   <list>
#> 28:      4.20       0   <list>
#> 29:      3.84       0   <list>
#> 30:      0.44       0   <list>
#> 31:      1.32       0   <list>
#> 32:      4.16       0   <list>
#> 33:      1.60       0   <list>
#> 34:      6.00       0   <list>
#> 35:      9.12       0   <list>
#> 36:      1.84       0   <list>
#> 37:      0.80       0   <list>
#> 38:      2.00       0   <list>
#> 39:      5.72       0   <list>
#> 40:      1.48       0   <list>
#> 41:      1.52       0   <list>
#> 42:     10.12       0   <list>
#> 43:      1.56       0   <list>
#> 44:     11.68       0   <list>
#> 45:      2.64       0   <list>
#> 46:      3.36       0   <list>
#> 47:      1.08       0   <list>
#> 48:      0.76       0   <list>
#> 49:      2.84       0   <list>
#> 50:      3.64       0   <list>
#> 51:     11.36       0   <list>
#> 52:      5.76       0   <list>
#> 53:     11.12       0   <list>
#> 54:      4.84       0   <list>
#> 55:     12.60       0   <list>
#> 56:     10.52       0   <list>
#> 57:     13.68       0   <list>
#> 58:     13.04       0   <list>
#> 59:      3.44       0   <list>
#> 60:      3.04       0   <list>
#> 61:      3.04       0   <list>
#> 62:     23.40       0   <list>
#> 63:      4.20       0   <list>
#> 64:      1.80       0   <list>
#> 65:     19.20       0   <list>
#> 66:     25.20       0   <list>
#> 67:     13.84       0   <list>
#> 68:      3.56       0   <list>
#> 69:      2.96       0   <list>
#> 70:      7.88       0   <list>
#>     mean_crps sd_crps R0_range
#> 
#> $rate_of_spread
#>                  type overall_little_r time_varying_r
#> 1: infection_upscaled           <list>   <data.table>
#> 
#> $raw_R0
#>            date         R sample crps window rt_type               type
#>   1: 2020-01-05 0.5111961      1 0.04      7 nowcast infection_upscaled
#>   2: 2020-01-05 0.9124742      1 0.04      7 nowcast infection_upscaled
#>   3: 2020-01-05 1.5000929      1 0.04      7 nowcast infection_upscaled
#>   4: 2020-01-05 2.1521715      1 0.04      7 nowcast infection_upscaled
#>   5: 2020-01-05 2.2019690      1 0.04      7 nowcast infection_upscaled
#>  ---                                                                   
#> 346: 2020-03-14 1.1638980      1 7.88      7 nowcast infection_upscaled
#> 347: 2020-03-14 1.1640007      1 7.88      7 nowcast infection_upscaled
#> 348: 2020-03-14 1.2003160      1 7.88      7 nowcast infection_upscaled
#> 349: 2020-03-14 1.2012230      1 7.88      7 nowcast infection_upscaled
#> 350: 2020-03-14 1.3065610      1 7.88      7 nowcast infection_upscaled
#>