Estimate time-varying measures and forecast

Estimate time-varying measures and forecast

epi_measures_pipeline(
  nowcast = NULL,
  serial_intervals = NULL,
  min_est_date = NULL,
  si_samples = NULL,
  rt_samples = NULL,
  rt_windows = 7,
  rate_window = 7,
  rt_prior = NULL,
  forecast_model = NULL,
  horizon = NULL,
  verbose = TRUE
)

Arguments

nowcast
A nowcast as produced by nowcast_pipeline

serial_intervals
A matrix with columns representing samples and rows representing the probability of the serial intervel being on that day.

min_est_date
Date to begin estimation.

si_samples
Numeric, the number of samples to take from the serial intervals supplied

rt_samples
Numeric, the number of samples to take from the estimated R distribution for each time point.

rt_windows
Numeric vector, windows over which to estimate time-varying R. The best performing window will be selected per serial interval sample by default (based on which window best forecasts current cases).

rate_window
Numeric, the window to use to estimate the rate of spread.

rt_prior
A list defining the reproduction number prior containing the mean (mean_prior) and standard deviation (std_prior)

forecast_model
An uninitialised bsts model passed to EpiSoon::forecast_rt to be used for forecasting future Rt values. An example of the required structure is: function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}.

horizon
Numeric, defaults to 0. The horizon over which to forecast Rts and cases.

verbose
Logical, defaults to TRUE. Should progress messages be shown.

Value

Examples