Estimate the time varying R0 - using EpiEstim

Estimate the time varying R0 - using EpiEstim

estimate_R0(
  cases = NULL,
  serial_intervals = NULL,
  rt_prior = NULL,
  window = NULL,
  si_samples = 100,
  rt_samples = 100
)

Arguments

cases
A dataframe containing a list of local cases with the following variables: date, cases, and import_status

serial_intervals
A matrix with columns representing samples and rows representing the probability of the serial intervel being on that day.

rt_prior
A list defining the reproduction number prior containing the mean (mean_prior) and standard deviation (std_prior)

window
Numeric, the window over which to estimate time-varying R

si_samples
Numeric, the number of samples to take from the serial intervals supplied

rt_samples
Numeric, the number of samples to take from the estimated R distribution for each time point.

Value

A tibble containing the date and summarised R estimte.

Examples